refactor(sync): LocalLibraryHostService 改名落地 + 按域拆 mixin（B4，2/2）
上一提交 9781b95279 只含 `git mv`（提交前 git add 链在已删除路径的 pathspec
上中断，commit 拿到的是 mv 时暂存的裸改名），本提交装全部内容；两条一起看，
合并时 squash。

改名：这个类从不依赖 AppModel（构造注入 db / packages / 回调，测试全用内存 DB），
名字里的 AppModel 是历史遗留；改成 LocalLibraryHostService（"本机库"的互联
host 实现），lib / test / packages 内 38 个文件 257 处引用同步改（docs/bugs 与
specs 历史记录不追改）。

拆分：2079 行主体拆成主库 + local_library_host_service/*.part.dart：
dictionaries / books / local_audio / audiobooks / videos / sync_state。这个类
90% 的方法是 7 个 host 接口的 @override 实现，extension 成员满足不了接口契约
（试过：88 条 override_on_non_overriding_member），所以按域单位是 **mixin**：
- 主库：`abstract class _LocalLibraryHostBase implements 7 接口 { 19 个抽象私有
  getter }`（从具体类的 final 字段机械生成）、具体类 `extends _Base with 7 个
  mixin`（字段加 @override 实现 getter）、`mixin _LocalLibraryHostShared on _Base`
  收三个跨域查询 helper（主合集归属一趟映射 / 可导出有声书键集）。
- 每个 part：`mixin _LocalLibraryHostX on _Base, _Shared { …@override 方法逐字… }`。
- private static helper 提到库顶层（mixin 体内看不到宿主类 static）：跨域的
  _assertSafeName / _existingFilePath 在主库，只被一个域用的在对应 part 尾部。
对包外零可见性变化、零行为改动。

守卫随扫描面搬家：新增 test/sync/local_library_host_service_source_corpus.dart
（主库 + 磁盘枚举全部 part）；media_cover_write_guard 的派生点 + 裸写点登记改
指 videos.part.dart（importVideo 派生 coversDir、_moveFileInto 裸写都在那）；
audio_reference_import_guard 指 audiobooks.part；video_delete_reclaim_entry_guard
的 deleteVideo 指 videos.part；fushi_library_collections_test 的 BUG-814 负向断言
改扫合并语料。server_lifecycle / host_book_import_wiring 两条守卫钉的是
app_model.dart 里的构造字面量，随改名 sed 一起换。

验证：flutter analyze 全量零问题；flutter test test/sync test/tools + 三条改过的
守卫 2877 全过。
