+++ fushi/lib/src/media/display_title.dart
@@ -30,7 +30,7 @@ import 'package:fushi_core/fushi_core.dart';
 /// 「母设备改过名的书，同步到子设备仍显示原书名」——因为改名只落本机偏好，
 /// 而三条出境通道全把它当设备设置。现在三条都通了，且**只搬显示名，绝不搬身份**：
 ///
-/// - 互联清单：`AppModelLibraryHostService.listBooks` 填 `RemoteBookInfo.displayTitle`
+/// - 互联清单：`LocalLibraryHostService.listBooks` 填 `RemoteBookInfo.displayTitle`
 ///   （additive wire 键，与 raw title 相同时不写；旧 host/旧 peer 天然回落 raw）。
 ///   消费恒走 `RemoteBookInfo.displayName`；`downloadId` / `bookKey` 仍是 raw。
 /// - 下载落地：peer 导入后把 host 的显示名写成本机 override（本机已有 override
