+++ fushi/lib/src/sync/local_library_host_service.dart
@@ -50,6 +50,69 @@ import 'package:fushi_dictionary/fushi_dictionary.dart';
 import 'package:fushi_audio/fushi_audio.dart' show AudiobookStorage;
 import 'package:path/path.dart' as p;
 
+part 'local_library_host_service/dictionaries.part.dart';
+part 'local_library_host_service/books.part.dart';
+part 'local_library_host_service/local_audio.part.dart';
+part 'local_library_host_service/audiobooks.part.dart';
+part 'local_library_host_service/videos.part.dart';
+part 'local_library_host_service/sync_state.part.dart';
+
+// ── 跨域共享的顶层 helper（原 LocalLibraryHostService 的 private static；mixin 体内看不到宿主类
+//    的 static，故提到库顶层——对包外零可见性变化）。
+
+/// 校验词典名称不含路径穿越字符。
+///
+/// 名称为空、或含 `/`、`\`、`..` 时抛 [ArgumentError]，
+/// 确保服务层自身也防御路径穿越，不依赖上层端点网关的单点防护。
+void _assertSafeName(String name) {
+  if (name.isEmpty ||
+      name.contains('/') ||
+      name.contains('\\') ||
+      name.contains('..')) {
+    throw ArgumentError.value(name, 'name', 'unsafe dictionary name');
+  }
+}
+
+String? _existingFilePath(String? path) {
+  if (path == null || path.isEmpty) return null;
+  return File(path).existsSync() ? path : null;
+}
+
+/// [LocalLibraryHostService] 的抽象骨架：宣告全部 host 接口，并把具体类的注入字段以抽象私有
+/// getter 暴露给各域 mixin（`mixin _X on _Base` 只能看到 `on` 类型声明的成员）。
+/// 库私有，包外不可见；接口成员由各域 mixin 逐字实现（B4 按域拆分）。
+abstract class _LocalLibraryHostBase
+    implements
+        FushiLibraryHostService,
+        DeletionTombstoneHost,
+        VideoDeletionHost,
+        VideoPlaybackSyncHost,
+        AudiobookDelayHost,
+        InterconnectServiceConfigHost,
+        InterconnectProfileHost {
+  FushiDatabase get _db;
+  Directory get _dictionaryResourceRoot;
+  SyncAssetPackageService get _packages;
+  Future<void> Function() get _refreshDictionaryCache;
+  Future<void> Function(Future<void> Function() body) get _runExclusive;
+  Future<bool> Function()? get _isProfileTransferEnabled;
+  Future<String> Function()? get _exportActiveProfileJson;
+  Future<String> Function(String json)? get _importProfileJson;
+  Future<String?> Function(File epubFile)? get _importBookFromFile;
+  Future<void> Function(EpubBookRow row)? get _cleanupBookOnDisk;
+  List<LocalAudioDbEntry> get _localAudioEntries;
+  Directory? get _localAudioStagingDir;
+  Future<void> Function(LocalAudioPackageContents)? get _onLocalAudioImported;
+  Directory? get _audioDatabaseRoot;
+  Future<void> Function(String displayName)? get _removeLocalAudioEntry;
+  String get _videoSubtitleLangCode;
+  Directory? get _uploadedVideoRoot;
+  Directory? get _videoCoversDirectory;
+  Future<String?> Function(
+      {required String videoPath,
+      required String bookUid})? get _extractVideoCover;
+}
+
 /// 用真实 Hibiki 库（Drift DB + [SyncAssetPackageService]）实现 host 库服务。
 ///
 /// 库变动经注入的 [runExclusive] 串行（生产传 `runExclusiveWithSync`），
@@ -69,16 +132,16 @@ import 'package:path/path.dart' as p;
 /// | [videoSubtitleLangCode] | 视频 sidecar 字幕匹配语言代码（P4-1）| AppModel 目标学习语言 |
 ///
 /// T2/T3 后续接线任务会在 AppModel 初始化时传入真实值。
-class AppModelLibraryHostService
-    implements
-        FushiLibraryHostService,
-        DeletionTombstoneHost,
-        VideoDeletionHost,
-        VideoPlaybackSyncHost,
-        AudiobookDelayHost,
-        InterconnectServiceConfigHost,
-        InterconnectProfileHost {
-  AppModelLibraryHostService({
+class LocalLibraryHostService extends _LocalLibraryHostBase
+    with
+        _LocalLibraryHostShared,
+        _LocalLibraryHostDictionaries,
+        _LocalLibraryHostBooks,
+        _LocalLibraryHostLocalAudio,
+        _LocalLibraryHostAudiobooks,
+        _LocalLibraryHostVideos,
+        _LocalLibraryHostSyncState {
+  LocalLibraryHostService({
     required FushiDatabase db,
     required Directory dictionaryResourceRoot,
     required SyncAssetPackageService packages,
@@ -121,236 +184,91 @@ class AppModelLibraryHostService
         _videoCoversDirectory = videoCoversDirectory,
         _extractVideoCover = extractVideoCover;
 
+  @override
   final FushiDatabase _db;
+  @override
   final Directory _dictionaryResourceRoot;
+  @override
   final SyncAssetPackageService _packages;
+  @override
   final Future<void> Function() _refreshDictionaryCache;
+  @override
   final Future<void> Function(Future<void> Function() body) _runExclusive;
 
   /// 互联「配置文件」（Profile）搬运的三个可选依赖（与本类其余可选能力同范式：
   /// 注入回调而不是把 ProfileRepository 的构造依赖整条拖进来）。生产由 AppModel
   /// 传入；未注入时开关判据恒为 false，端点对外表现为「host 关着」。
+  @override
   final Future<bool> Function()? _isProfileTransferEnabled;
+  @override
   final Future<String> Function()? _exportActiveProfileJson;
+  @override
   final Future<String> Function(String json)? _importProfileJson;
-
-  @override
-  Future<bool> isInterconnectProfileTransferEnabled() async {
-    // 三者缺一即视为不可用：没有导出/导入回调时开着开关也无从服务。
-    final Future<bool> Function()? enabled = _isProfileTransferEnabled;
-    if (enabled == null) return false;
-    if (_exportActiveProfileJson == null || _importProfileJson == null) {
-      return false;
-    }
-    return enabled();
-  }
-
-  /// 导出**也**要串行：激活 Profile 的导出会先 `snapshotCurrentSettings()`，
-  /// 那是一次写库 —— 一个远端 GET 就能和正在跑的备份恢复 / 集合同步交错。
-  @override
-  Future<String> exportInterconnectProfile() async {
-    final Future<String> Function()? export = _exportActiveProfileJson;
-    if (export == null) {
-      throw UnsupportedError('profile export not wired on this host');
-    }
-    late final String json;
-    await _runExclusive(() async {
-      json = await export();
-    });
-    return json;
-  }
-
-  /// 导入必须串行：底下的 `importProfileFromJson` 是跨三个 await 的
-  /// check-then-act（算唯一名 → 建 Profile → 写设置）。两个并发对端（或一个对端
-  /// 重试）会算出**同一个名字**，而且在建行与写设置之间存在一个「有名字、零设置」
-  /// 的半成品 Profile 对本机 UI 可见。client 侧的 `_busy` 只挡本机重复点击，挡不
-  /// 住并发对端 —— 串行锁得由 host 侧兜。
-  @override
-  Future<String> importInterconnectProfile(String json) async {
-    final Future<String> Function(String json)? import = _importProfileJson;
-    if (import == null) {
-      throw UnsupportedError('profile import not wired on this host');
-    }
-    late final String name;
-    await _runExclusive(() async {
-      name = await import(json);
-    });
-    return name;
-  }
-
-  @override
-  Future<InterconnectServiceConfigSnapshot>
-      getInterconnectServiceConfig() async {
-    return InterconnectServiceConfigSnapshot.fromPreferences(
-      await _db.getAllPrefs(),
-    );
-  }
 
   /// 书籍导入回调（可选；null 时 importBook 抛 [UnsupportedError]）。
   /// 生产传 `(f) => EpubImporter.importFromPath(db: db, filePath: f.path, fileName: p.basename(f.path))`。
   /// BUG-1503：返回落地后的真实 bookKey（重名加 `(2)` 后缀会改变派生键），
   /// 供 [importBook] 把推送方随行的显示名挂到正确的书上。
+  @override
   final Future<String?> Function(File epubFile)? _importBookFromFile;
 
   /// 书籍磁盘清理回调（可选；null 时只执行 DB 删除，跳过 AudiobookStorage/SrtBook 清理）。
   /// 生产传 ReaderFushiSource 实例的磁盘清理部分。
+  @override
   final Future<void> Function(EpubBookRow row)? _cleanupBookOnDisk;
 
   // ── 本地音频（T3.1）──────────────────────────────────────────────────────
 
   /// 当前已注册的本地音频来源列表。生产传 AppModel.localAudioEntries。
+  @override
   final List<LocalAudioDbEntry> _localAudioEntries;
 
   /// importLocalAudio 解包用临时目录。null 时用 Directory.systemTemp。
+  @override
   final Directory? _localAudioStagingDir;
 
   /// 本地音频包解包后的注册回调（可选；null 时 importLocalAudio 抛 [UnsupportedError]）。
   /// 生产传 AppModel.importSyncedLocalAudioDb。
+  @override
   final Future<void> Function(LocalAudioPackageContents)? _onLocalAudioImported;
 
   // ── 有声书（T3.1）────────────────────────────────────────────────────────
 
   /// importAudiobook 音频文件落盘根目录（可选；null 时 importAudiobook 抛 [UnsupportedError]）。
   /// 生产传 AppModel 的 audioDatabaseRoot。
+  @override
   final Directory? _audioDatabaseRoot;
 
   /// deleteLocalAudio 回调（可选；null 时 deleteLocalAudio 仅做名称校验，静默跳过删除）。
   /// 生产传按 displayName 从 LocalAudioManager 移除条目的回调（T3.4 接线）。
+  @override
   final Future<void> Function(String displayName)? _removeLocalAudioEntry;
 
   // ── 视频（P4-1）──────────────────────────────────────────────────────────────
 
   /// 视频 sidecar 字幕匹配的目标语言代码（默认 'ja'）。
   /// 生产传 JapaneseLanguage.instance.languageCode（P4 接线任务完成后注入真实值）。
+  @override
   final String _videoSubtitleLangCode;
 
   /// client→host 上传视频的落盘根目录（可选；null 时 [importVideo] 抛 [UnsupportedError]）。
   /// 生产传 `<documents>/remote_videos`（[AppPaths.remoteVideosDirectory] 同目录，与
   /// client 下载远端视频落点一致）。
+  @override
   final Directory? _uploadedVideoRoot;
+  @override
   final Directory? _videoCoversDirectory;
 
   /// 上传视频后的封面抽取回调（可选、best-effort；null 时上传的视频无封面占位）。
   /// 生产传 `extractVideoCover`（桌面 ffmpeg 抽帧；移动端无 ffmpeg 返 null 留空占位）。
+  @override
   final Future<String?> Function(
       {required String videoPath, required String bookUid})? _extractVideoCover;
+}
 
-  static const String _dictionaryAssetSuffix = '.fushidict';
-
-  /// 校验词典名称不含路径穿越字符。
-  ///
-  /// 名称为空、或含 `/`、`\`、`..` 时抛 [ArgumentError]，
-  /// 确保服务层自身也防御路径穿越，不依赖上层端点网关的单点防护。
-  static void _assertSafeName(String name) {
-    if (name.isEmpty ||
-        name.contains('/') ||
-        name.contains('\\') ||
-        name.contains('..')) {
-      throw ArgumentError.value(name, 'name', 'unsafe dictionary name');
-    }
-  }
-
-  static EpubBookRow? _findBookByTitleOrKey(
-    List<EpubBookRow> rows,
-    String titleOrBookKey,
-  ) =>
-      rows.cast<EpubBookRow?>().firstWhere(
-            (EpubBookRow? r) =>
-                r!.bookKey == titleOrBookKey || r.title == titleOrBookKey,
-            orElse: () => null,
-          );
-
-  static String? _existingFilePath(String? path) {
-    if (path == null || path.isEmpty) return null;
-    return File(path).existsSync() ? path : null;
-  }
-
-  /// host 当前实时词典清单（从 DictionaryMeta 表读）。
-  @override
-  Future<List<RemoteDictionaryInfo>> listDictionaries() async {
-    final List<DictionaryMetaRow> rows = await _db.getAllDictionaryMetadata();
-    return <RemoteDictionaryInfo>[
-      for (final DictionaryMetaRow r in rows)
-        RemoteDictionaryInfo(name: r.name, type: r.type),
-    ];
-  }
-
-  /// 即时把名为 [name] 的实时词典打包成临时 .fushidict 文件，返回该文件。
-  /// 调用方负责删除返回的临时文件（及其父临时目录）。词典不存在抛 [StateError]。
-  /// 名称含路径穿越字符时抛 [ArgumentError]。
-  @override
-  Future<File> exportDictionary(String name) async {
-    _assertSafeName(name);
-    final List<DictionaryMetaRow> rows = await _db.getAllDictionaryMetadata();
-    final bool exists = rows.any((DictionaryMetaRow r) => r.name == name);
-    if (!exists) throw StateError('dictionary not found: $name');
-
-    final Directory tmpDir =
-        Directory.systemTemp.createTempSync('fushi_dict_export');
-    final File out = File(p.join(tmpDir.path, '$name$_dictionaryAssetSuffix'));
-    await _packages.exportDictionaryPackage(
-      dictionaryName: name,
-      dictionaryResourceRoot: _dictionaryResourceRoot,
-      outputFile: out,
-    );
-    return out;
-  }
-
-  /// 把 [packageFile]（.fushidict）导入 host 实时库（幂等：同名覆盖资源 + upsert 元数据）。
-  @override
-  Future<void> importDictionary(File packageFile) async {
-    await _runExclusive(() async {
-      // 同名覆盖会往既有词典目录里写 blobs.bin / hash.table，而这些文件正被引擎
-      // MapViewOfFile 映射着 —— Windows 拒绝以写方式打开，覆盖导入直接失败
-      // （BUG-1756）。收尾的 _refreshDictionaryCache 会把引擎重新加载回来。
-      //
-      // 装回必须走 finally：releaseAllMappings 之后引擎是空的，而
-      // importDictionaryPackage 会因包损坏 / 磁盘满 / 权限抛出。写在 try 之后就被
-      // 跳过 —— host 引擎永久空转，所有互联对端查词返回空，直到 host 重启。
-      FushiDicts.releaseAllMappings();
-      try {
-        await _packages.importDictionaryPackage(
-          packageFile: packageFile,
-          dictionaryResourceRoot: _dictionaryResourceRoot,
-        );
-      } finally {
-        await _refreshDictionaryCache();
-      }
-    });
-  }
-
-  /// 从 host 实时库删除名为 [name] 的词典（DB 元数据 + 资源目录）。
-  /// 名称含路径穿越字符时抛 [ArgumentError]。
-  @override
-  Future<void> deleteDictionary(String name) async {
-    _assertSafeName(name);
-    await _runExclusive(() async {
-      await _db.deleteDictionaryMeta(name);
-      // 顺序不可交换（BUG-1756）：先 refresh —— 它 loadFromDb 把 host 刚改过的 DB
-      // 同步进 Dart 侧 cache，再据此重载引擎，于是被删的那本立刻从引擎里掉出去、
-      // 它的 mmap view 被释放。反过来写（先删目录）在 Windows 上必抛
-      // ERROR_USER_MAPPED_FILE，且 refresh 永远执行不到。
-      //
-      // 这一次 refresh 不能省成「交给原语的 reloadEngine」：目录不存在时原语直接
-      // 返回、不碰引擎，cache 就会停在「已删的词典还在」的旧状态上。
-      await _refreshDictionaryCache();
-      await deleteDictionaryDirectory(
-        Directory(p.join(_dictionaryResourceRoot.path, name)),
-        reloadEngine: _refreshDictionaryCache,
-      );
-    });
-  }
-
-  // ── 书籍 ─────────────────────────────────────────────────────────────────
-
-  /// bookKey → 标签名列表 的一趟映射（TODO-1165，避免逐条 N+1 查询）。
-  /// 复用 DB 层 SQL 过滤的批查（review-reuse-2：内层 map 的 key 即标签名，
-  /// 别再全表拉 assignments 手工 join）。
-  Future<Map<String, List<String>>> _tagNamesByBookKey() async =>
-      (await _db.allBookTagAddedAtByName()).map(
-          (String key, Map<String, int> byName) =>
-              MapEntry(key, byName.keys.toList()));
-
+/// 跨域共享的查询 helper（书 / 视频 / 有声书都用）：主合集归属一趟映射、可导出有声书键集。
+/// 方法逐字搬自 LocalLibraryHostService；各域 mixin 以 `on _LocalLibraryHostBase, _LocalLibraryHostShared` 取用。
+mixin _LocalLibraryHostShared on _LocalLibraryHostBase {
   /// `'<mediaType>|<entryKey>'` → 该条目的**主合集归属**（多端库联合视图 §2.3
   /// 任务5.1）的一趟映射。归属跟随 [FushiDatabase.getPrimaryCollectionIdByEntry] 的
   /// 「最小 collectionId」折叠语义：一条目属多合集时只带它折进的那一张，与库网格
@@ -409,468 +327,6 @@ class AppModelLibraryHostService
     return (membership: out, collectionByEntry: rowByEntry);
   }
 
-  /// videoBookUid → 标签名列表 的一趟映射（TODO-1165）。
-  Future<Map<String, List<String>>> _tagNamesByVideoUid() async =>
-      (await _db.allVideoTagAddedAtByName()).map(
-          (String key, Map<String, int> byName) =>
-              MapEntry(key, byName.keys.toList()));
-
-  /// host 当前书库清单（从 EpubBooks 表读）。
-  /// [RemoteBookInfo.hasContent] 为 true 当且仅当该书存在可导出的 EPUB 根目录。
-  @override
-  Future<List<RemoteBookInfo>> listBooks() async {
-    final List<EpubBookRow> rows = await _db.getAllEpubBooks();
-    // 该书是否已配「可经 live-sync 导出」的有声书。判据必须与 [listAudiobooks] /
-    // [exportAudiobook] 完全同源——仅 Audiobooks 行还不够，导出格式需 srtBookUid，
-    // 故只有同时具备 SrtBooks 行的 bookKey 才算可下载（TODO-778）。否则 EPUB 对齐
-    // 有声书（有 Audiobook 无 SrtBook）会亮徽章 + 可点下载，但 exportAudiobook
-    // 抛 StateError → 服务端 404。
-    final Set<String> audiobookKeys = await _srtBackedAudiobookKeys();
-    final Map<String, List<String>> tagsByBookKey = await _tagNamesByBookKey();
-    // tags 稳健档：host 为每本书带上标签 LWW 时钟（名→加入戳）+ 移除墓碑，供 client
-    // mergeRemoteBookTags 传播 host 侧的删除/改名、防复活（旧 client 忽略这些键、按
-    // tags 名单只增，向后兼容）。批量一趟查（旧实现逐书 2 次查询，大库清单端点 O(N)
-    // 次 DB 往返）。
-    final Map<String, Map<String, int>> tagAddedAtByKey =
-        await _db.allBookTagAddedAtByName();
-    final Map<String, Map<String, int>> tagTombByKey =
-        await _db.allTagTombstonesByName(MediaKind.epub);
-    final Map<String, RemoteCollectionMembership> membership =
-        await _primaryCollectionMembership();
-    // BUG-1488：host 上用户改过的书名（`preferences` 的 override 覆盖层）随清单
-    // 下发，否则 peer 永远只看得到 raw `epub_books.title`。一趟 prefs 读。
-    final Map<String, OverrideTitleEntry> overrideTitles =
-        await _overrideTitleByBookKey();
-    // 阅读进度内联（首页仪表盘「继续」的互联数据源）：percent 与本地首页同源同算
-    // （MediaItems.position/duration），最近阅读时刻取 reader_positions.updatedAt。
-    // 各一趟批查；旧 client 忽略这两个 additive 字段。
-    final Map<String, ({int percent, int updatedAtMs})> progressByKey =
-        await _bookProgressByKey();
-    // BUG-812：srt-backed 有声书（同 bookKey 既有 EpubBooks 又有 SrtBooks 行）加入合集
-    // 时以 **`srt|<uid>`** 存进成员表（本地书架把它当 SRT 卡渲染、经 srt|uid 折叠），
-    // 而非 `epub|<bookKey>`。互联 client 把这类书作为 EPUB 占位卡收下，只查 `epub|bookKey`
-    // 会 miss → RemoteBookInfo.collection=null → 有声书不折进合集。这里补一趟
-    // bookKey→srtUid 映射，供下面按 srt 成员键兜底查归属，让有声书 EPUB 卡也带上
-    // collection（client 经现有注入循环折进合集，与本地书架对称、不依赖后台 sweep）。
-    final Map<String, String> srtUidByBookKey = <String, String>{
-      for (final SrtBookRow s in await _db.getAllSrtBooks())
-        if (s.bookKey.isNotEmpty) s.bookKey: s.uid,
-    };
-    return rows.map((EpubBookRow r) {
-      // EPUB 行的 coverPath 是 EPUB 内部相对 href，必须拼 extractDir 才是磁盘真
-      // 路径；直接 _existingFilePath(相对href) 恒 false → 远端书卡没封面（#4）。
-      final String? coverPath = resolveEpubCoverFilePath(
-        extractDir: r.extractDir,
-        coverPath: r.coverPath,
-      );
-      final BookFormat format = BookFormat.parseOrEpub(r.format);
-      return RemoteBookInfo(
-        title: r.title,
-        // BUG-1488：只在真改过名时非 null（toJson 亦只在与 title 不同时写键）。
-        displayTitle: overrideTitles[r.bookKey]?.title,
-        // BUG-1502：改名时刻随行，让 client 端做 LWW 而不是 insert-if-absent。
-        displayTitleAt: overrideTitles[r.bookKey]?.updatedAt ?? 0,
-        bookKey: r.bookKey,
-        // hasContent（EPUB 内容树可导出）按 format 门控（互联完整支持批次）：由
-        // EPUB 转化来的漫画行 extractDir 里仍留着 EPUB 解压树，旧判据会把它当可
-        // 下载 EPUB 打包——把整套页图 + manga.json 塞进 zip，client 落地成一本
-        // 夹带全部页图的「文字书」、漫画身份静默丢失（坏包）。漫画内容走
-        // hasMangaContent + 漫画包通道。
-        hasContent: format == BookFormat.epub &&
-            resolveExtractedEpubRoot(r.extractDir) != null,
-        format: format.dbValue,
-        hasMangaContent: format == BookFormat.manga &&
-            File(p.join(r.extractDir, kMangaPackageMarker)).existsSync(),
-        mangaReadingMode: r.mangaReadingMode,
-        hasEmbeddedCover: coverPath != null,
-        coverPath: coverPath,
-        hasAudiobook: audiobookKeys.contains(r.bookKey),
-        tags: tagsByBookKey[r.bookKey] ?? const <String>[],
-        tagsAddedAt: tagAddedAtByKey[r.bookKey] ?? const <String, int>{},
-        tagTombstones: tagTombByKey[r.bookKey] ?? const <String, int>{},
-        // 合集成员键（v83）：epub 成员行 entryKey = 本机 epub_books.uid，书侧组键
-        // 用 r.uid（§2.3 任务5.1 的 wire 面貌仍是 bookKey，与本地键域无关）。
-        // epub|bookKey 兜底：sync 落地的透传行（书当时还没下载）在下一轮 apply
-        // 收敛前仍以对端 bookKey 为键，窗口期内按 bookKey 也查一把，归属不闪断。
-        // srt-backed 有声书兜底：该书以 srt|uid 入合集时，epub 两键都 miss，
-        // 回退查 srt|uid（BUG-812）。散卡（三键都无）= null。
-        collection: membership[MediaKind.epub.compositeKey(r.uid)] ??
-            membership[MediaKind.epub.compositeKey(r.bookKey)] ??
-            (srtUidByBookKey[r.bookKey] != null
-                ? membership[
-                    MediaKind.srt.compositeKey(srtUidByBookKey[r.bookKey]!)]
-                : null),
-        progressPercent: progressByKey[r.bookKey]?.percent ?? 0,
-        progressUpdatedAtMs: progressByKey[r.bookKey]?.updatedAtMs ?? 0,
-        // BUG-1119：EpubBooks 行都是可下载 EPUB，显式标 epub（srt-backed 有声书
-        // 的 EPUB 卡语义仍是 epub——与本地 _bookMediaKind 按 hoshi://book/ 身份判
-        // epub 一致，勿标成 srt 造成两端同书异 kind）。standalone SRT 书（身份
-        // hoshi://srtbook/<uid>，无 EpubBooks 行）今天不在本清单，进清单是独立
-        // follow-up。
-        kind: MediaKind.epub,
-      );
-    }).toList();
-  }
-
-  /// 批查「bookKey → 用户自定义显示名」（BUG-1488）。
-  ///
-  /// 改名不改 `epub_books.title`（标题派生 bookKey 是跨端身份，改列 = 十来张子表
-  /// 连坐改键），而是往 `preferences` 写一行覆盖，key 形如
-  /// `src:reader_fushi:override_title://fushi://book/<bookKey>`。三段前缀分别由
-  /// [dbSourcePrefKey] / [MediaSource.overrideTitleKeyFor] /
-  /// [ReaderFushiSource.mediaIdentifierFor] 各自的真相源拼出，本层零硬编码字符串。
-  ///
-  /// 只认**规范**键形态：BUG-1317 之前的旧键（源键出现两次）由读取期回退在 host
-  /// 自己的书架上就地重写成规范键，而改名动作本身恒写规范键，所以旧形态只可能
-  /// 属于「BUG-1317 之前改的名 + 此后从未在本机显示过」的书，可忽略。
-  /// bookKey → (host 上的显示名, 它的 LWW 毫秒戳)。
-  ///
-  /// BUG-1502：戳来自 `preferences.updated_at`，随清单一起下发，让 client 端能对
-  /// 「本机也改过名」的书做 last-write-wins（此前只能 insert-if-absent，host 的
-  /// **第二次**改名永远传不过去）。实现共享给推书方向（BUG-1503），见
-  /// [readOverrideTitlesByBookKey]。
-  Future<Map<String, OverrideTitleEntry>> _overrideTitleByBookKey() =>
-      readOverrideTitlesByBookKey(_db);
-
-  static final RegExp _fushiBookKeyPattern = RegExp(r'^fushi://book/(.+)$');
-
-  /// 批查全库书籍阅读进度「bookKey → (percent 0..100, 最近阅读毫秒戳)」。
-  ///
-  /// percent 与本地首页仪表盘「继续」完全同算：`MediaItems.position/duration`
-  /// （书的 MediaItem.mediaIdentifier = `hoshi://book/<bookKey>`）；时刻取
-  /// `reader_positions.updatedAt`。两趟全表读，无逐书查询。
-  Future<Map<String, ({int percent, int updatedAtMs})>>
-      _bookProgressByKey() async {
-    // v82：reader_positions 键 = 书 uid，wire/mediaId 面貌仍是 bookKey——
-    // 经 epub_books 反查（uid → bookKey）后出 wire。
-    final Map<String, String> bookKeyByUid = <String, String>{
-      for (final EpubBookRow b in await _db.getAllEpubBooks())
-        if (b.uid.isNotEmpty) b.uid: b.bookKey,
-    };
-    final Map<String, int> updatedAtByKey = <String, int>{
-      for (final ReaderPositionRow r in await _db.getAllReaderPositions())
-        if (bookKeyByUid[r.bookUid] case final String key) key: r.updatedAt,
-    };
-    final Map<String, ({int percent, int updatedAtMs})> out =
-        <String, ({int percent, int updatedAtMs})>{};
-    for (final MediaOpenHistoryRow m in await _db.getAllMediaOpenHistory()) {
-      final RegExpMatch? match = _fushiBookKeyPattern.firstMatch(m.mediaId);
-      if (match == null) continue;
-      final String bookKey = match.group(1)!;
-      if (m.duration <= 0 || m.position <= 0) continue;
-      out[bookKey] = (
-        percent: ((m.position / m.duration) * 100).clamp(0, 100).round(),
-        updatedAtMs: updatedAtByKey[bookKey] ?? 0,
-      );
-    }
-    return out;
-  }
-
-  /// host 最近 [limit] 条活动事件（新首页 Activity 面板互联数据源）。
-  @override
-  Future<List<RemoteActivityEvent>> listActivityEvents(
-      {int limit = 100}) async {
-    // v92：活动流唯一数据源是统一事实面（legacy 活动行 ∪ 段 ∪ 游玩会话合成行），
-    // 与本机首页同一份；否则 client 看不到 host 在 v92 之后的任何阅读 / 观看。
-    final List<ActivityEventRow> rows =
-        (await loadStatFacts(_db, activityLimit: limit)).activityRows;
-    return <RemoteActivityEvent>[
-      for (final ActivityEventRow r in rows)
-        RemoteActivityEvent(
-          eventType: r.eventType,
-          mediaType: r.mediaType,
-          title: r.title,
-          dateKey: r.dateKey,
-          timestampMs: r.timestampMs,
-          mediaKey: r.mediaKey,
-          durationMs: r.durationMs,
-          charsDelta: r.charsDelta,
-        ),
-    ];
-  }
-
-  /// 即时把书名为 [title] 的书 extractDir 重打包成 .epub 临时文件，返回该文件。
-  /// 调用方负责删除返回的临时文件（及其父临时目录）。
-  /// [title] 含路径穿越字符时抛 [ArgumentError]；
-  /// 书不存在或 extractDir 为空/不存在时抛 [StateError]。
-  @override
-  Future<File> exportBook(String title) async {
-    _assertSafeName(title);
-    final List<EpubBookRow> rows = await _db.getAllEpubBooks();
-    final EpubBookRow? row = _findBookByTitleOrKey(rows, title);
-    if (row == null) {
-      throw StateError('book not found: $title');
-    }
-    final BookFormat format = BookFormat.parseOrEpub(row.format);
-    // 漫画（互联完整支持批次）：包 = 书目录整树 zip（manga.json 标记 + 页图），
-    // 与 EPUB 包同端点、导入侧内容嗅探分流。扩展名仍 .epub（端点/tmp 命名契约
-    // 不变，内容即真相）。
-    if (format == BookFormat.manga) {
-      final Directory tmpDir =
-          Directory.systemTemp.createTempSync('hibiki_book_export');
-      final File out = File(p.join(tmpDir.path, '${row.bookKey}.epub'));
-      final bool ok = await repackageMangaBook(row.extractDir, out.path);
-      if (!ok) {
-        throw StateError('manga book has no exportable content: $title');
-      }
-      return out;
-    }
-    if (format != BookFormat.epub ||
-        resolveExtractedEpubRoot(row.extractDir) == null) {
-      // PDF（无互联内容通道）/ EPUB 树缺失：与旧行为一致抛 StateError → 404。
-      throw StateError('book has no exportable EPUB root: $title');
-    }
-
-    final Directory tmpDir =
-        Directory.systemTemp.createTempSync('hibiki_book_export');
-    // 文件名用 title 但扩展名用 .epub，保证重导入时 fileName 是合法 epub 名。
-    final String safeBasename = '${row.bookKey}.epub';
-    final File out = File(p.join(tmpDir.path, safeBasename));
-    final bool ok = await repackageExtractedEpub(row.extractDir, out.path);
-    if (!ok) {
-      throw StateError('repackage produced no output for book: $title');
-    }
-    return out;
-  }
-
-  /// 把 [epubFile] 导入 host 书库。
-  ///
-  /// 生产使用时需在构造器传入 [importBookFromFile] 回调
-  /// （例如 `(f) => EpubImporter.importFromPath(db: db, filePath: f.path, fileName: p.basename(f.path))`）。
-  /// 回调为 null 时抛 [UnsupportedError]。
-  ///
-  /// BUG-1503：回调返回**落地后的真实 bookKey**（`EpubImporter.importFromPath` 的
-  /// 返回值），显示名才有地方挂——它不能由 [displayTitle] 或 URL 里的 title 推，
-  /// 重名时 importer 会加 `(2)` 后缀，派生键随之不同。
-  @override
-  Future<void> importBook(
-    File epubFile, {
-    String? displayTitle,
-    int displayTitleAt = 0,
-  }) async {
-    final Future<String?> Function(File)? importer = _importBookFromFile;
-    if (importer == null) {
-      throw UnsupportedError(
-        'importBook requires importBookFromFile callback to be provided',
-      );
-    }
-    String? bookKey;
-    await _runExclusive(() async {
-      bookKey = await importer(epubFile);
-    });
-    await _adoptPushedDisplayTitle(
-      bookKey: bookKey,
-      displayTitle: displayTitle,
-      displayTitleAt: displayTitleAt,
-    );
-  }
-
-  /// 把推送方随书带来的显示名落成 host 本机的 override，last-write-wins
-  /// （BUG-1503 + BUG-1502）。
-  ///
-  /// 与 client 端下载后的采纳（`_adoptRemoteBookDisplayTitle`）和备份合并
-  /// （`BackupMergeEngine._mergeOverrideTitlePrefs`）共用同一条裁决规则：严格更新
-  /// 才覆盖、平局保留本机、本机没有该行则无条件采纳。三条通道语义不一致就等于没修。
-  ///
-  /// 走 `MediaSource.adoptOverrideTitleIfNewer` 而不是裸写 DB：host 是个正在跑的
-  /// app，`MediaSource` 有一层内存偏好缓存，只写 DB 的话 host 书架会一直显示旧名
-  /// 直到重启（而且 `getPreference` 的 miss 会把 null 反写进缓存）。
-  Future<void> _adoptPushedDisplayTitle({
-    required String? bookKey,
-    required String? displayTitle,
-    required int displayTitleAt,
-  }) async {
-    if (bookKey == null ||
-        bookKey.isEmpty ||
-        displayTitle == null ||
-        displayTitle.isEmpty) {
-      return;
-    }
-    try {
-      final ReaderFushiSource source = ReaderFushiSource.instance;
-      await source.adoptOverrideTitleIfNewer(
-        item: source.overrideTitleMediaItemForBookKey(bookKey),
-        title: displayTitle,
-        updatedAt: displayTitleAt,
-      );
-    } catch (e, stack) {
-      // 书已经入库了——显示名没落上不该把整个 PUT 变成 HTTP 500。
-      ErrorLogService.instance
-          .log('AppModelLibraryHostService.adoptPushedDisplayTitle', e, stack);
-    }
-  }
-
-  /// 从 host 书库删除书名为 [title] 的书（DB 行 + 磁盘目录）。
-  /// [title] 含路径穿越字符时抛 [ArgumentError]。
-  @override
-  Future<void> deleteBook(String title) async {
-    _assertSafeName(title);
-    await _runExclusive(() async {
-      final List<EpubBookRow> rows = await _db.getAllEpubBooks();
-      final EpubBookRow? row = _findBookByTitleOrKey(rows, title);
-      if (row == null) return; // 幂等：不存在则静默跳过
-
-      // 先让注入的磁盘清理回调运行（AudiobookStorage / SrtBook 等 DB 行外资源），
-      // 在 DB deleteEpubBook 事务之前拿到 row 数据（事务后 row 即消失）。
-      await _cleanupBookOnDisk?.call(row);
-
-      // DB 事务：删除 EpubBooks 行及其所有关联行（readerPositions / bookmarks /
-      // srtBooks / audioCues / audiobooks）。见 HBK-AUDIT-041。
-      // TODO-1195 part B：用户删书记墓碑，避免旧备份合并导入时复活。
-      await _db.deleteEpubBook(row.bookKey, tombstone: true);
-
-      // extractDir 磁盘目录：DB 删除后再清理（与 reader_fushi_source 同顺序）。
-      if (row.extractDir.isNotEmpty) {
-        final Directory dir = Directory(row.extractDir);
-        if (dir.existsSync()) await dir.delete(recursive: true);
-      }
-    });
-  }
-
-  /// 读 host 端书 [bookKey] 的阅读进度（TODO-767）。直读 host 自己的
-  /// `reader_positions` 表（与 host 本地阅读该书时同一真相源）；无记录返回
-  /// [RemoteBookProgress.empty]。
-  @override
-  Future<RemoteBookProgress> getBookProgress(String bookKey) async {
-    // v82：wire 键 bookKey → 本地子表键 uid 换算；书不在库视同无记录。
-    final EpubBookRow? book = await _db.getEpubBook(bookKey);
-    if (book == null || book.uid.isEmpty) return RemoteBookProgress.empty;
-    final ReaderPositionRow? row = await _db.getReaderPosition(book.uid);
-    if (row == null) return RemoteBookProgress.empty;
-    return RemoteBookProgress(
-      sectionIndex: row.sectionIndex,
-      normCharOffset: row.normCharOffset,
-      charOffset: row.charOffset,
-      updatedAtMs: row.updatedAt,
-    );
-  }
-
-  /// 把 client 上报的书 [bookKey] 进度写入 host 自己的 `reader_positions`
-  /// （TODO-767）。
-  ///
-  /// 冲突解决「取较新时间戳」（[resolveBookProgressSync]）：仅当 [progress] 严格
-  /// 新于 host 已存时间戳才覆盖，避免旧设备滞后上报回退新进度。胜出方等于 host 已存
-  /// 进度时 no-op（不写库）。负 normCharOffset clamp 0。
-  @override
-  Future<void> putBookProgress(
-    String bookKey,
-    RemoteBookProgress progress,
-  ) async {
-    // host 书库不存在该 bookKey → no-op，不写孤儿 `reader_positions` 行。
-    // （reader_positions 无外键也无 GC，任意 client 上报任意 bookKey 都会落库；
-    // 之后 host 若导入同名 sanitize bookKey 的书，恢复时会取到来自别处设备、host
-    // 从没读过的陈旧位置 = 进度污染。与视频 `updateVideoBookPosition`「UPDATE
-    // 不存在即 no-op」语义对齐。syncContent 开时 client 独有书已先经
-    // `_syncBooksContentLive` importBook 推成 host 书，故正常同步不被此闸门误挡。）
-    final EpubBookRow? hostBook = await _db.getEpubBook(bookKey);
-    if (hostBook == null || hostBook.uid.isEmpty) return;
-    final RemoteBookProgress current = await getBookProgress(bookKey);
-    final RemoteBookProgress incoming = RemoteBookProgress(
-      sectionIndex: progress.sectionIndex < 0 ? 0 : progress.sectionIndex,
-      normCharOffset: progress.normCharOffset < 0 ? 0 : progress.normCharOffset,
-      charOffset: progress.charOffset,
-      updatedAtMs: progress.updatedAtMs,
-    );
-    final RemoteBookProgress winner =
-        resolveBookProgressSync(local: current, remote: incoming);
-    if (winner.sectionIndex == current.sectionIndex &&
-        winner.normCharOffset == current.normCharOffset &&
-        winner.charOffset == current.charOffset &&
-        winner.updatedAtMs == current.updatedAtMs) {
-      return; // host 已存更新或相等，no-op。
-    }
-    await _runExclusive(() async {
-      await _db.upsertReaderPosition(ReaderPositionsCompanion(
-        bookUid: Value(hostBook.uid),
-        sectionIndex: Value(winner.sectionIndex),
-        normCharOffset: Value(winner.normCharOffset),
-        charOffset: Value(winner.charOffset),
-        updatedAt: Value(winner.updatedAtMs),
-      ));
-    });
-  }
-
-  // ── 本地音频（T3.1）────────────────────────────────────────────────────────
-
-  /// host 当前本地音频来源清单（从注入的 [_localAudioEntries] 取 displayName）。
-  @override
-  Future<List<RemoteLocalAudioInfo>> listLocalAudio() async {
-    return <RemoteLocalAudioInfo>[
-      for (final LocalAudioDbEntry e in _localAudioEntries)
-        RemoteLocalAudioInfo(displayName: e.displayName),
-    ];
-  }
-
-  /// 即时把 displayName 为 [displayName] 的本地音频库打包成临时文件，返回该文件。
-  /// 调用方负责删除返回的临时文件（及其父临时目录）。
-  /// [displayName] 含路径穿越字符时抛 [ArgumentError]；
-  /// 找不到该来源或其 DB 文件不存在时抛 [StateError]。
-  @override
-  Future<File> exportLocalAudio(String displayName) async {
-    _assertSafeName(displayName);
-    final LocalAudioDbEntry? entry =
-        _localAudioEntries.cast<LocalAudioDbEntry?>().firstWhere(
-              (LocalAudioDbEntry? e) => e!.displayName == displayName,
-              orElse: () => null,
-            );
-    if (entry == null) {
-      throw StateError('local audio not found: $displayName');
-    }
-    final File dbFile = File(entry.path);
-    if (!dbFile.existsSync()) {
-      throw StateError('local audio DB file not found: ${entry.path}');
-    }
-
-    final Directory tmpDir =
-        Directory.systemTemp.createTempSync('hibiki_local_audio_export');
-    final File out = File(p.join(tmpDir.path, '$displayName.fushiaudiolib'));
-    await _packages.exportLocalAudioPackage(
-      displayName: entry.displayName,
-      enabled: entry.enabled,
-      sources: entry.sources,
-      dbFile: dbFile,
-      outputFile: out,
-    );
-    return out;
-  }
-
-  /// 把本地音频包文件导入 host（解包 + 注册）。
-  /// 需要在构造器传入 [onLocalAudioImported] 回调；回调为 null 时抛 [UnsupportedError]。
-  @override
-  Future<void> importLocalAudio(File packageFile) async {
-    final Future<void> Function(LocalAudioPackageContents)? callback =
-        _onLocalAudioImported;
-    if (callback == null) {
-      throw UnsupportedError(
-        'importLocalAudio requires onLocalAudioImported callback to be provided',
-      );
-    }
-    await _runExclusive(() async {
-      final Directory stagingDir =
-          _localAudioStagingDir ?? Directory.systemTemp;
-      final LocalAudioPackageContents contents =
-          await _packages.importLocalAudioPackage(
-        packageFile: packageFile,
-        stagingDir: stagingDir,
-      );
-      await callback(contents);
-    });
-  }
-
-  /// 从 host 删除 displayName 为 [displayName] 的本地音频来源。
-  ///
-  /// 注：本地音频来源的注册信息存于 Preferences（不在 Drift DB），删除需经
-  /// [_removeLocalAudioEntry] 回调，生产由 AppModel 注入。回调为 null 时静默
-  /// 跳过实际删除（等同 no-op，保持幂等）。
-  /// [displayName] 含路径穿越字符时抛 [ArgumentError]。
-  @override
-  Future<void> deleteLocalAudio(String displayName) async {
-    _assertSafeName(displayName);
-    final Future<void> Function(String)? remover = _removeLocalAudioEntry;
-    if (remover == null) return; // 回调未注入：静默跳过（幂等）
-    await _runExclusive(() => remover(displayName));
-  }
-
   // ── 有声书包（T3.1）────────────────────────────────────────────────────────
 
   /// 既有 Audiobooks 行又有 SrtBooks 行的 bookKey 集合——即真正可经 live-sync
@@ -887,1193 +343,5 @@ class AppModelLibraryHostService
       }
     }
     return keys;
-  }
-
-  /// host 当前可导出的有声书清单。
-  ///
-  /// 两类：
-  /// - **srt-backed**（既有 Audiobooks 又有 SrtBooks 行）：身份键 = bookKey（不变）。
-  /// - **纯 SRT（standalone）有声书**（SrtBooks 行 bookKey 为空、无 Audiobooks 行）：
-  ///   身份键 = uid。旧枚举只遍历 Audiobooks 表，完全遗漏这类书 → 无法跨设备下载。
-  @override
-  Future<List<RemoteAudiobookInfo>> listAudiobooks() async {
-    final List<AudiobookRow> rows = await _db.getAllAudiobooks();
-    final List<RemoteAudiobookInfo> result = <RemoteAudiobookInfo>[];
-    final Set<String> emittedUids = <String>{};
-    // 清单内联断点 + 调轴（互联完整支持批次）：与视频清单同范式，sweep/下载回填
-    // 免逐本 GET（旧实现 N 本书 = N 次网络往返）。一趟 prefs 批读。
-    final Map<String, String> allPrefs = await _db.getAllPrefs();
-    RemoteAudiobookInfo build(String bookKey, String uid, String? title) {
-      final String identity = bookKey.isNotEmpty ? bookKey : uid;
-      return RemoteAudiobookInfo(
-        bookKey: bookKey,
-        uid: uid,
-        title: title,
-        positionMs: PrefCodec.decode<int>(
-            allPrefs[audiobookPositionPrefKey(identity)] ?? '', 0),
-        positionUpdatedAtMs: PrefCodec.decode<int>(
-            allPrefs[audiobookPositionAtPrefKey(identity)] ?? '', 0),
-        delayMs: PrefCodec.decode<int>(
-            allPrefs[audiobookDelayPrefKey(identity)] ?? '', 0),
-        delayUpdatedAtMs: PrefCodec.decode<int>(
-            allPrefs[audiobookDelayAtPrefKey(identity)] ?? '', 0),
-      );
-    }
-
-    for (final AudiobookRow r in rows) {
-      final SrtBookRow? srt = await _db.getSrtBookByBookKey(r.bookKey);
-      if (srt == null) continue;
-      result.add(build(r.bookKey, srt.uid, srt.title));
-      emittedUids.add(srt.uid);
-    }
-    // 纯 SRT（standalone）有声书：bookKey 为空、不落 Audiobooks 行，身份 = uid。
-    final List<SrtBookRow> srtRows = await _db.getAllSrtBooks();
-    for (final SrtBookRow srt in srtRows) {
-      if (srt.bookKey.isNotEmpty) continue; // srt-backed 已在上面枚举
-      if (!emittedUids.add(srt.uid)) continue; // 去重（防重复 uid）
-      result.add(build('', srt.uid, srt.title));
-    }
-    return result;
-  }
-
-  /// 读 host 端有声书 [identity] 的调轴（[AudiobookDelayHost]）。值键是既有
-  /// `audiobook_delay_` pref（旧数据无戳记 0，被任何带戳对端值盖过）。
-  @override
-  Future<({int delayMs, int updatedAtMs})> getAudiobookDelay(
-      String identity) async {
-    final int delay =
-        await _db.getPrefTyped<int>(audiobookDelayPrefKey(identity), 0);
-    final int at =
-        await _db.getPrefTyped<int>(audiobookDelayAtPrefKey(identity), 0);
-    return (delayMs: delay, updatedAtMs: at);
-  }
-
-  /// 把 client 上报的有声书 [identity] 调轴写入 host（[AudiobookDelayHost]）。
-  /// 存在性闸门 / clamp / 未来戳截断 / LWW 与视频 [putVideoDelay] 同纪律；
-  /// 有声书调轴本地读取就是这对 prefs，写入即对 host 本机播放生效（无行写穿）。
-  @override
-  Future<void> putAudiobookDelay(
-      String identity, int delayMs, int updatedAtMs) async {
-    if (!await audiobookExists(identity)) return;
-    final int clamped =
-        delayMs.clamp(-kVideoSubtitleDelayLimitMs, kVideoSubtitleDelayLimitMs);
-    final int nowCapMs = DateTime.now().millisecondsSinceEpoch + 5 * 60 * 1000;
-    final int cappedAt = updatedAtMs > nowCapMs ? nowCapMs : updatedAtMs;
-    final ({int delayMs, int updatedAtMs}) current =
-        await getAudiobookDelay(identity);
-    final ({int delayMs, int updatedAtMs}) winner = resolveDelayLww(
-      aDelayMs: current.delayMs,
-      aUpdatedAtMs: current.updatedAtMs,
-      bDelayMs: clamped,
-      bUpdatedAtMs: cappedAt,
-    );
-    if (winner.updatedAtMs == current.updatedAtMs &&
-        winner.delayMs == current.delayMs) {
-      return; // host 已存更新或相等，no-op。
-    }
-    await _db.setPrefTyped<int>(
-        audiobookDelayPrefKey(identity), winner.delayMs);
-    await _db.setPrefTyped<int>(
-        audiobookDelayAtPrefKey(identity), winner.updatedAtMs);
-  }
-
-  /// 即时把身份键为 [identity] 的有声书打包成临时文件，返回该文件。
-  /// 调用方负责删除返回的临时文件（及其父临时目录）。
-  ///
-  /// [identity] 解析：先按 bookKey 查 Audiobooks（srt-backed），命中即打含 audiobook
-  /// 段的包；否则按 uid 查 SrtBooks（纯 SRT standalone），命中即打无 audiobook 段的
-  /// 纯 SRT 包。两者都查不到抛 [StateError]。含路径穿越字符时抛 [ArgumentError]。
-  @override
-  Future<File> exportAudiobook(String identity) async {
-    _assertSafeName(identity);
-
-    // srt-backed：identity = bookKey，Audiobooks 行 + SrtBooks 行齐备。
-    final AudiobookRow? ab = await _db.getAudiobookByBookKey(identity);
-    if (ab != null) {
-      final SrtBookRow? srt = await _db.getSrtBookByBookKey(identity);
-      if (srt == null) {
-        throw StateError('srtBook not found for bookKey: $identity');
-      }
-      return _packAudiobook(
-        identity: identity,
-        srtBookUid: srt.uid,
-        bookKey: identity,
-      );
-    }
-
-    // 纯 SRT（standalone）：identity = uid，无 Audiobooks 行，bookKey 空。
-    final SrtBookRow? srtStandalone = await _db.getSrtBookByUid(identity);
-    if (srtStandalone != null) {
-      return _packAudiobook(
-        identity: identity,
-        srtBookUid: identity,
-        bookKey: null, // 无 Audiobooks 行 → 打纯 SRT 包
-      );
-    }
-
-    throw StateError('audiobook not found for identity: $identity');
-  }
-
-  /// 打包成 `<identity>.fushiaudio` 临时文件（srt-backed 传 [bookKey]；纯 SRT 传
-  /// null，包管线据此省略 audiobook 段、cue 走 uid 命名空间）。
-  Future<File> _packAudiobook({
-    required String identity,
-    required String srtBookUid,
-    required String? bookKey,
-  }) async {
-    final Directory tmpDir =
-        Directory.systemTemp.createTempSync('hibiki_audiobook_export');
-    final File out = File(p.join(tmpDir.path, '$identity.fushiaudio'));
-    await _packages.exportAudioDatabasePackage(
-      srtBookUid: srtBookUid,
-      outputFile: out,
-      bookKey: bookKey,
-    );
-    return out;
-  }
-
-  /// 廉价判断 host 库是否存在 bookKey 为 [bookKey] 的有声书（BUG-471a）：仅一次
-  /// `Audiobooks` 行查询，不触发 [exportAudiobook] 的整包打包 zip I/O。与
-  /// [putAudiobookPosition] 自身用的存在性闸门同一查询。
-  @override
-  Future<bool> audiobookExists(String identity) async {
-    _assertSafeName(identity);
-    if (await _db.getAudiobookByBookKey(identity) != null) return true;
-    // 纯 SRT standalone：无 Audiobooks 行，按 uid 查 SrtBooks。
-    return await _db.getSrtBookByUid(identity) != null;
-  }
-
-  /// 把有声书包文件导入 host（解包写 DB + 音频文件）。
-  /// 需要在构造器传入 [audioDatabaseRoot]；为 null 时抛 [UnsupportedError]。
-  @override
-  Future<void> importAudiobook(File packageFile,
-      {String? bookKeyOverride}) async {
-    final Directory? root = _audioDatabaseRoot;
-    if (root == null) {
-      throw UnsupportedError(
-        'importAudiobook requires audioDatabaseRoot to be provided',
-      );
-    }
-    await _runExclusive(() async {
-      await _packages.importAudioDatabasePackage(
-        packageFile: packageFile,
-        audioDatabaseRoot: root,
-        bookKeyOverride: bookKeyOverride,
-      );
-    });
-  }
-
-  /// 从 host 删除 bookKey 为 [bookKey] 的有声书（Audiobooks/SrtBooks/AudioCues 行
-  /// + 磁盘音频目录）。[bookKey] 含路径穿越字符时抛 [ArgumentError]；
-  /// 不存在则静默跳过（幂等）。
-  @override
-  Future<void> deleteAudiobook(String bookKey) async {
-    _assertSafeName(bookKey);
-    await _runExclusive(() async {
-      final AudiobookRow? ab = await _db.getAudiobookByBookKey(bookKey);
-      if (ab != null) {
-        // 先取 audioRoot，再删 DB 行（磁盘清理在 DB 删除后，同 deleteBook 顺序）。
-        final String? audioRoot = ab.audioRoot;
-
-        // 删除 SrtBooks 行（按 bookKey），其关联的 SrtBook 级别 audioCues 由事务处理。
-        // getSrtBookByBookKey 先拿 uid，再用 deleteSrtBookByUid 级联删 audioCue 行。
-        final SrtBookRow? srt = await _db.getSrtBookByBookKey(bookKey);
-        if (srt != null) {
-          await _db.deleteSrtBookByUid(srt.uid);
-        }
-
-        // 删除 Audiobooks 行（及其 audioCues 级联，via deleteAudiobookByBookKey）。
-        await _db.deleteAudiobookByBookKey(bookKey);
-
-        await _deleteAudioRootIfPersisted(audioRoot);
-        return;
-      }
-
-      // 纯 SRT standalone：identity = uid，无 Audiobooks 行。按 uid 删 SrtBooks 行
-      // （级联 uid 命名空间的 audioCues）+ 其持久音频目录。
-      final SrtBookRow? srt = await _db.getSrtBookByUid(bookKey);
-      if (srt == null) return; // 幂等：都不存在则静默跳过
-      final String? audioRoot = srt.audioRoot;
-      await _db.deleteSrtBookByUid(srt.uid);
-      await _deleteAudioRootIfPersisted(audioRoot);
-    });
-  }
-
-  /// 删除 [audioRoot] 磁盘目录，但仅当它在 app 内部持久根（<appDoc>/audiobooks）下
-  /// —— 绝不递归删用户「引用导入」的原始外部目录（TODO-935 ①A）。
-  Future<void> _deleteAudioRootIfPersisted(String? audioRoot) async {
-    if (audioRoot == null || audioRoot.isEmpty) return;
-    final String persistRoot = await AudiobookStorage.audiobooksRootDir();
-    final bool referenced = AudiobookStorage.isReferencedPath(
-      filePath: audioRoot,
-      persistRoot: persistRoot,
-    );
-    if (!referenced) {
-      final Directory dir = Directory(audioRoot);
-      if (dir.existsSync()) await dir.delete(recursive: true);
-    }
-  }
-
-  /// 读 host 端有声书 [bookKey] 的播放断点（BUG-471）。真相源是
-  /// `audiobook_pos_<bookKey>` + `audiobook_pos_at_<bookKey>` prefs（host 本机播放
-  /// 与远端 resume 路径统一写此键空间，见 [AudiobookRepository.updatePositionMs]）。
-  ///
-  /// 向后兼容：旧数据只写位置不写时间戳，缺时间戳时记 0，被任何带时间戳的对端进度
-  /// 在 [resolvePositionLww] 中盖过——既能读出旧本机播放位置，又不让无时间戳
-  /// 旧值盖过更新的对端进度。
-  @override
-  Future<({int positionMs, int updatedAtMs})> getAudiobookPosition(
-    String bookKey,
-  ) async {
-    final int pos =
-        await _db.getPrefTyped<int>(audiobookPositionPrefKey(bookKey), 0);
-    final int at =
-        await _db.getPrefTyped<int>(audiobookPositionAtPrefKey(bookKey), 0);
-    return (positionMs: pos, updatedAtMs: at);
-  }
-
-  /// 把 client 上报的有声书 [bookKey] 断点写入 host（BUG-471）。
-  ///
-  /// 存在性闸门：host 无该 bookKey 的 Audiobooks 行 → no-op，不写孤儿
-  /// `audiobook_pos_` pref（与视频 [putVideoPosition]「视频不存在不写脏」、书
-  /// [putBookProgress]「书不存在不写孤儿行」同语义）。
-  ///
-  /// 冲突解决「取较新时间戳」（[resolvePositionLww]）：仅当 [updatedAtMs]
-  /// 严格新于 host 已存时间戳才覆盖。负位置 clamp 0。
-  @override
-  Future<void> putAudiobookPosition(
-    String bookKey,
-    int positionMs,
-    int updatedAtMs,
-  ) async {
-    // host 库不存在该有声书 → no-op（防任意 client 上报任意 key 写脏 prefs）。
-    // srt-backed 按 bookKey 查 Audiobooks；纯 SRT standalone 按 uid 查 SrtBooks。
-    // 进度 pref key = audiobook_pos_<identity>：standalone 的 identity=uid 恰为
-    // SrtBook 进度键，故写穿即写到正确命名空间。
-    if (await _db.getAudiobookByBookKey(bookKey) == null &&
-        await _db.getSrtBookByUid(bookKey) == null) {
-      return;
-    }
-    final ({int positionMs, int updatedAtMs}) current =
-        await getAudiobookPosition(bookKey);
-    final ({int positionMs, int updatedAtMs}) winner = resolvePositionLww(
-      localPositionMs: current.positionMs,
-      localUpdatedAtMs: current.updatedAtMs,
-      remotePositionMs: positionMs < 0 ? 0 : positionMs,
-      remoteUpdatedAtMs: updatedAtMs,
-    );
-    if (winner.updatedAtMs == current.updatedAtMs &&
-        winner.positionMs == current.positionMs) {
-      return; // host 已存更新或相等，no-op。
-    }
-    await _db.setPrefTyped<int>(
-        audiobookPositionPrefKey(bookKey), winner.positionMs);
-    await _db.setPrefTyped<int>(
-        audiobookPositionAtPrefKey(bookKey), winner.updatedAtMs);
-  }
-
-  // ── 视频（P4-1，只读）────────────────────────────────────────────────────────
-
-  /// host 当前视频清单（从 VideoBooks 表读，按 importedAt DESC 排序）。
-  ///
-  /// [sizeBytes] 取 videoPath 对应文件的大小（stat），文件不存在时为 null。
-  /// [durationMs] 目前恒为 null（DB 无 duration 列，后续由 ffprobe/libmpv 填充）。
-  /// [hasSubtitle] 当前视频文件旁能找到外挂字幕时为 true。
-  @override
-  Future<List<RemoteVideoInfo>> listVideos() async {
-    final List<VideoBookRow> rows = await _db.allVideoBooks();
-    // 按 importedAt 降序（null 排最后）
-    rows.sort((VideoBookRow a, VideoBookRow b) {
-      final int? ta = a.importedAt;
-      final int? tb = b.importedAt;
-      if (ta == null && tb == null) return 0;
-      if (ta == null) return 1;
-      if (tb == null) return -1;
-      return tb.compareTo(ta);
-    });
-
-    final Map<String, List<String>> tagsByVideoUid =
-        await _tagNamesByVideoUid();
-    final ({
-      Map<String, RemoteCollectionMembership> membership,
-      Map<String, MediaCollectionRow> collectionByEntry,
-    }) collections = await _primaryCollectionData();
-    final Map<String, RemoteCollectionMembership> membership =
-        collections.membership;
-    // 批量预取：位置 prefs / 标签 LWW 时钟 / 移除墓碑各一趟查询，sidecar 同目录只扫
-    // 一次。旧实现逐行 2~3 次 prefs 读 + 1 次行重查 + 2 次标签查询 + 1 次全目录
-    // listSync——500 行清单一次 ≈ 2500 次 DB 往返 + 500 次目录扫描，且封面端点每张
-    // 封面重跑整份清单时按 N² 放大（见 [videoCoverPath]）。
-    final Map<String, String> allPrefs = await _db.getAllPrefs();
-    final Map<String, Map<String, int>> tagAddedAtByUid =
-        await _db.allVideoTagAddedAtByName();
-    final Map<String, Map<String, int>> tagTombByUid =
-        await _db.allTagTombstonesByName(MediaKind.video);
-    final Map<String, List<String>?> sidecarDirCache =
-        <String, List<String>?>{};
-    final List<RemoteVideoInfo> videos = <RemoteVideoInfo>[];
-    for (final VideoBookRow row in rows) {
-      videos.add(_videoInfoFromRow(
-        row,
-        tags: tagsByVideoUid[row.bookUid] ?? const <String>[],
-        // 合集成员键：video 条目 mediaType='video'、entryKey=bookUid（§2.3 任务5.1）。
-        collection: membership[MediaKind.video.compositeKey(row.bookUid)],
-        collectionRow: collections
-            .collectionByEntry[MediaKind.video.compositeKey(row.bookUid)],
-        prefs: allPrefs,
-        tagsAddedAt: tagAddedAtByUid[row.bookUid] ?? const <String, int>{},
-        tagTombstones: tagTombByUid[row.bookUid] ?? const <String, int>{},
-        sidecarDirCache: sidecarDirCache,
-      ));
-    }
-    return videos;
-  }
-
-  /// 按 [id] 单查视频封面磁盘路径（封面端点专用，一次 DB 单行查询 + stat；绝不
-  /// materialize 整份 [listVideos]——旧封面路径每张封面重跑全量清单是 O(N²) 主犯）。
-  @override
-  Future<String?> videoCoverPath(String id) async {
-    final VideoBookRow? row = await _db.getVideoBookByBookUid(id);
-    if (row == null) return null;
-    return _existingFilePath(row.coverPath);
-  }
-
-  /// 按 [id]（downloadId：bookKey 或 title）单查书封面磁盘路径（对称
-  /// [videoCoverPath]）。优先按 bookKey 单行查；未命中（旧 client 用 title 当
-  /// downloadId）回退全表按 title/bookKey 匹配——仍只是一次 DB 查询 + 几次 stat，
-  /// 没有旧 listBooks 的逐书标签/有声书/合集重活。
-  @override
-  Future<String?> bookCoverPath(String id) async {
-    EpubBookRow? row = await _db.getEpubBook(id);
-    row ??= _findBookByTitleOrKey(await _db.getAllEpubBooks(), id);
-    if (row == null) return null;
-    return resolveEpubCoverFilePath(
-      extractDir: row.extractDir,
-      coverPath: row.coverPath,
-    );
-  }
-
-  /// 列出 [dir] 下的文件名（仅文件，不含子目录）；目录不存在/读取失败返回 null。
-  /// [listVideos] 用它配合每次调用内的目录缓存，同目录 500 个视频只扫一次。
-  static List<String>? _listDirFileNames(String dir) {
-    final Directory directory = Directory(dir);
-    try {
-      if (!directory.existsSync()) return null;
-      return directory
-          .listSync(followLinks: false)
-          .whereType<File>()
-          .map((File f) => p.basename(f.path))
-          .toList();
-    } on FileSystemException {
-      return null;
-    }
-  }
-
-  /// 构建单条 [RemoteVideoInfo]（内部辅助，纯同步：所有 DB 数据均由 [listVideos]
-  /// 批量预取后经参数注入，仅保留文件 stat 与目录缓存内的 sidecar 匹配）。
-  RemoteVideoInfo _videoInfoFromRow(
-    VideoBookRow row, {
-    required Map<String, String> prefs,
-    required Map<String, int> tagsAddedAt,
-    required Map<String, int> tagTombstones,
-    required Map<String, List<String>?> sidecarDirCache,
-    List<String> tags = const <String>[],
-    RemoteCollectionMembership? collection,
-    MediaCollectionRow? collectionRow,
-  }) {
-    final String videoPath = row.videoPath;
-    int? sizeBytes;
-    bool hasSubtitle = false;
-    String? subtitleFileName;
-
-    if (videoPath.isNotEmpty) {
-      final File f = File(videoPath);
-      if (f.existsSync()) {
-        try {
-          sizeBytes = f.lengthSync();
-        } catch (_) {
-          // stat 失败：保守返回 null
-        }
-        // 检查外挂字幕 sidecar（廉价：目录 listing 每目录只扫一次 + 纯字符串匹配）。
-        final String dir = p.dirname(videoPath);
-        final List<String>? dirFiles =
-            sidecarDirCache.putIfAbsent(dir, () => _listDirFileNames(dir));
-        final String? picked = dirFiles == null
-            ? null
-            : pickSidecar(
-                p.basenameWithoutExtension(videoPath),
-                dirFiles,
-                langCode: _videoSubtitleLangCode,
-              );
-        if (picked != null) {
-          hasSubtitle = true;
-          subtitleFileName = picked;
-        }
-        // BUG-814：列表端点**不做**内嵌字幕轨 ffmpeg 探测。旧实现在此逐视频串行
-        // spawn `ffmpeg -i`（每项超时基线 60s、大文件到 1200s、无缓存、每次 GET 全量
-        // 重跑），大库（如 511 个视频）轻易超过 client 的 15s listTimeout → 远端视频
-        // 判空 → 手机整页空。内嵌轨是**播放时**才需要的信息，已由 `/streamurl` 端点
-        // (`fushi_sync_server.dart` `_embeddedSubtitleTracksForRequest`) 在拉流时按需
-        // 探测并下发（client 唯一消费者 video_fushi_page 读的是 streamurl 响应，列表
-        // 的 embeddedSubtitleTracks 零消费）。故此处保持 embeddedSubtitleTracks 为空、
-        // hasSubtitle 只反映廉价的外挂 sidecar——列表变纯 DB/stat 读，与 listBooks 对称、
-        // 毫秒返回。
-      }
-    }
-
-    final String? coverPath = _existingFilePath(row.coverPath);
-    // TODO-653: 把 host 端记录的播放断点带进清单条目，供 client 跨设备恢复。
-    // 语义与 [getVideoPosition]\(id, episodeIndex: 0\) 完全一致：prefs 断点（本机/
-    // 远端播放统一键）与旧 `VideoBooks.lastPositionMs`（时间戳 0）取较新——只是行
-    // 已在手、prefs 已批量预取，不再逐行发查询。
-    // 系列级播放偏好（schema v52 同系列共享）：合集成员的无戳基底 = 系列级 ?? 本集
-    // row 值（与本机播放页 [effectiveSeriesDelayMs] 同一决议）。此前只读 row，host
-    // 在合集里调的轴远端永远看不到。
-    // 播放偏好带戳状态（播放偏好同步泛化批）：语义与 [getVideoPlayback] 完全一致
-    // ——无戳基底（系列级 ?? row）与带戳 prefs 逐字段严格较新者胜；prefs 已批量
-    // 预取，不逐行发查询。
-    String? nullableStr(String? raw) {
-      if (raw == null) return null;
-      final String decoded = PrefCodec.decode<String>(raw, '');
-      return decoded.isEmpty ? null : decoded;
-    }
-
-    final VideoPlaybackSyncState playback = VideoPlaybackSyncState.merge(
-      VideoPlaybackSyncState(
-        delayMs:
-            effectiveSeriesDelayMs(collectionRow?.subtitleDelayMs, row.delayMs),
-        audioTrackId: effectiveSeriesAudioTrackId(
-            collectionRow?.audioTrackId, row.audioTrackId),
-        secondarySubtitleSource: row.secondarySubtitleSource,
-        secondaryDelayMs: effectiveSeriesSecondaryDelayMs(
-            collectionRow?.secondarySubtitleDelayMs, row.secondaryDelayMs),
-      ),
-      VideoPlaybackSyncState(
-        delayMs: PrefCodec.decode<int>(
-            prefs[videoRemoteDelayPrefKey(row.bookUid)] ?? '', 0),
-        delayAt: PrefCodec.decode<int>(
-            prefs[videoRemoteDelayAtPrefKey(row.bookUid)] ?? '', 0),
-        audioTrackId:
-            nullableStr(prefs[videoRemoteAudioTrackPrefKey(row.bookUid)]),
-        audioTrackAt: PrefCodec.decode<int>(
-            prefs[videoRemoteAudioTrackAtPrefKey(row.bookUid)] ?? '', 0),
-        secondarySubtitleSource: nullableStr(
-            prefs[videoRemoteSecondarySubtitlePrefKey(row.bookUid)]),
-        secondarySubtitleAt: PrefCodec.decode<int>(
-            prefs[videoRemoteSecondarySubtitleAtPrefKey(row.bookUid)] ?? '', 0),
-        secondaryDelayMs: int.tryParse(
-            nullableStr(prefs[videoRemoteSecondaryDelayPrefKey(row.bookUid)]) ??
-                ''),
-        secondaryDelayAt: PrefCodec.decode<int>(
-            prefs[videoRemoteSecondaryDelayAtPrefKey(row.bookUid)] ?? '', 0),
-      ),
-    );
-    final ({int positionMs, int updatedAtMs}) progress = resolvePositionLww(
-      localPositionMs: PrefCodec.decode<int>(
-          prefs[videoRemotePositionPrefKey(row.bookUid)] ?? '', 0),
-      localUpdatedAtMs: PrefCodec.decode<int>(
-          prefs[videoRemotePositionAtPrefKey(row.bookUid)] ?? '', 0),
-      remotePositionMs: row.lastPositionMs,
-      remoteUpdatedAtMs: 0,
-    );
-    // TODO-885: 解析 playlistJson → 远端剧集（只 index+title，绝不带 host path）。
-    final List<RemoteVideoEpisode> episodes = _episodesFromRow(row);
-    final int currentEpisode = episodes.length > 1
-        ? row.currentEpisode.clamp(0, episodes.length - 1)
-        : 0;
-    return RemoteVideoInfo(
-      id: row.bookUid,
-      title: row.title,
-      sizeBytes: sizeBytes,
-      hasSubtitle: hasSubtitle,
-      subtitleFileName: subtitleFileName,
-      embeddedSubtitleTracks: const <RemoteVideoEmbeddedSubtitleTrack>[],
-      // durationMs: 暂为 null，DB 无此列（后续接线任务填充）
-      hasCover: coverPath != null,
-      coverPath: coverPath,
-      positionMs: progress.positionMs,
-      positionUpdatedAtMs: progress.updatedAtMs,
-      // 播放偏好随清单下发（BUG-996 调轴起步，播放偏好同步泛化批扩展为全字段）：
-      // client 起播据带戳字段做逐字段 LWW 决议，sweep 免逐视频 GET。
-      delayMs: playback.delayMs,
-      delayUpdatedAtMs: playback.delayAt,
-      audioTrackId: playback.audioTrackId,
-      audioTrackUpdatedAtMs: playback.audioTrackAt,
-      secondarySubtitleSource: playback.secondarySubtitleSource,
-      secondarySubtitleUpdatedAtMs: playback.secondarySubtitleAt,
-      secondaryDelayMs: playback.secondaryDelayMs,
-      secondaryDelayUpdatedAtMs: playback.secondaryDelayAt,
-      // 看完标记下发（client 剧集面板角标；此前远端集无口径恒无标记）。
-      completedAt: row.completedAt?.millisecondsSinceEpoch,
-      episodes: episodes,
-      currentEpisode: currentEpisode,
-      tags: tags,
-      // tags 稳健档：带上标签 LWW 时钟 + 移除墓碑，供 client mergeRemoteVideoTags 传播
-      // host 侧删除/改名、防复活（旧 client 忽略、按 tags 名单只增）。
-      tagsAddedAt: tagsAddedAt,
-      tagTombstones: tagTombstones,
-      collection: collection,
-      // 入库时刻下发：client 的「最近添加」行与合集组间序都按它排；不下发就只能
-      // 给远端占位造假 importedAt，远端条目结构上进不了「最近添加」。
-      importedAt: row.importedAt,
-    );
-  }
-
-  /// 把 [row] 的 `playlistJson` 解析成远端剧集列表（TODO-885）。坏 JSON / 单视频
-  /// （≤1 集）返回空列表 = 单视频语义（向后兼容）。**只取 index+title**，host 端
-  /// 文件 path 留在 host（client 用 episodeIndex 反查），绝不进 [RemoteVideoEpisode]。
-  List<RemoteVideoEpisode> _episodesFromRow(VideoBookRow row) {
-    final List<PlaylistEntry> entries = _parsePlaylistEntries(row.playlistJson);
-    if (entries.length <= 1) return const <RemoteVideoEpisode>[];
-    return <RemoteVideoEpisode>[
-      for (int i = 0; i < entries.length; i++)
-        RemoteVideoEpisode(index: i, title: entries[i].title),
-    ];
-  }
-
-  /// 纯解析 `playlistJson` 为 [PlaylistEntry] 列表（坏 JSON 返回空）。host 端按集反查
-  /// 文件 path 用（[_resolveEpisodeVideoPath]）。
-  List<PlaylistEntry> _parsePlaylistEntries(String? playlistJson) {
-    if (playlistJson == null || playlistJson.isEmpty) {
-      return const <PlaylistEntry>[];
-    }
-    try {
-      final dynamic decoded = jsonDecode(playlistJson);
-      if (decoded is! List) return const <PlaylistEntry>[];
-      return <PlaylistEntry>[
-        for (final dynamic e in decoded)
-          if (e is Map) PlaylistEntry.fromJson(e.cast<String, dynamic>()),
-      ];
-    } catch (_) {
-      return const <PlaylistEntry>[];
-    }
-  }
-
-  /// 按 (bookUid=[id], [episodeIndex]) 从 host DB 反查该集真实视频文件路径（TODO-885）。
-  ///
-  /// **DB-only 安全契约**：path 永远来自 host 自己 `playlistJson` 解析，绝不接受外部
-  /// 传入。[episodeIndex]<=0 或非播放列表时回退 `videoPath`（当前选中集 / 单视频）。
-  /// 越界 [episodeIndex] 返回 null（安全拒绝）。
-  Future<String?> _resolveEpisodeVideoPath(String id, int episodeIndex) async {
-    if (episodeIndex < 0) return null; // 非法下标安全拒绝。
-    final VideoBookRow? row = await _db.getVideoBookByBookUid(id);
-    if (row == null) return null;
-    // 当前集 / 单视频（episodeIndex==0）：用 row.videoPath，等价旧行为。
-    if (episodeIndex == 0) {
-      return row.videoPath.isEmpty ? null : row.videoPath;
-    }
-    // 播放列表按集：DB 解析 playlistJson，越界安全拒绝。
-    final List<PlaylistEntry> entries = _parsePlaylistEntries(row.playlistJson);
-    if (episodeIndex >= entries.length) return null;
-    final String path = entries[episodeIndex].path;
-    return path.isEmpty ? null : path;
-  }
-
-  /// 按 [id]（即 `VideoBooks.bookUid`）反查真实视频文件。
-  ///
-  /// **只查 DB**，不接受外部文件路径。文件不存在或 id 未知时返回 null。
-  @override
-  Future<File?> resolveVideoFile(String id, {int episodeIndex = 0}) async {
-    final String? path = await _resolveEpisodeVideoPath(id, episodeIndex);
-    if (path == null || path.isEmpty) return null;
-    final File f = File(path);
-    return f.existsSync() ? f : null;
-  }
-
-  /// 按 [id] 查找对应视频的外挂字幕文件（sidecar）。
-  ///
-  /// 用 [langCode] 优先匹配带语言标记的字幕（如 `.ja.srt`）；内封字幕不在此列。
-  /// 找不到外挂字幕或视频未知时返回 null。
-  @override
-  Future<File?> resolveVideoSubtitle(
-    String id, {
-    String langCode = '',
-    int episodeIndex = 0,
-  }) async {
-    final String? videoPath = await _resolveEpisodeVideoPath(id, episodeIndex);
-    if (videoPath == null || videoPath.isEmpty) return null;
-    final String effectiveLangCode =
-        langCode.isEmpty ? _videoSubtitleLangCode : langCode;
-    final String? subPath =
-        findSidecarSubtitle(videoPath, langCode: effectiveLangCode);
-    if (subPath == null) return null;
-    final File f = File(subPath);
-    return f.existsSync() ? f : null;
-  }
-
-  /// BUG-1004：host 端本地裁 mining 句子音频（见抽象声明）。用 [resolveVideoFile] 反查真实
-  /// 本地文件后调 [extractAudioSegmentViaFfmpeg]（本地路径、不经网络/TLS——绕开 client
-  /// ffmpeg 抓 host 自签 https/token 流的整类失败）。裁到独立临时目录，产物返回给调用方，
-  /// 调用方读完删该目录；失败清理临时目录并返回 null。
-  @override
-  Future<File?> clipVideoAudio(
-    String id, {
-    required int startMs,
-    required int endMs,
-    int episodeIndex = 0,
-    int? audioStreamIndex,
-    int? audioStreamCount,
-    int audioChannels = 1,
-    String audioBitrate = '64k',
-  }) async {
-    if (endMs <= startMs) return null;
-    final File? file = await resolveVideoFile(id, episodeIndex: episodeIndex);
-    if (file == null) return null;
-    final Directory tmp =
-        Directory.systemTemp.createTempSync('hibiki_clip_audio');
-    final String out = p.join(tmp.path, 'clip.aac');
-    final String? result = await extractAudioSegmentViaFfmpeg(
-      inputPath: file.path,
-      startMs: startMs,
-      endMs: endMs,
-      outputPath: out,
-      audioStreamIndex: audioStreamIndex,
-      audioStreamCount: audioStreamCount,
-      audioChannels: audioChannels,
-      audioBitrate: audioBitrate,
-    );
-    if (result == null) {
-      try {
-        tmp.deleteSync(recursive: true);
-      } catch (_) {
-        // best-effort：临时目录清理失败不影响返回 null（裁切已失败）。
-      }
-      return null;
-    }
-    return File(result);
-  }
-
-  /// 读 host 端 [id] 视频的播放断点（TODO-653 / TODO-816 断点②）。
-  ///
-  /// 真相源是 `video_remote_position_<bookUid>` + `video_remote_position_at_<bookUid>`
-  /// prefs（host 本机播放与远端 resume 路径统一写此键空间，见 video_fushi_page
-  /// `_persistPosition` / `_persistRemotePosition`）。
-  ///
-  /// 向后兼容：TODO-816 之前 host 本机播放只写 `VideoBooks.lastPositionMs`、不写 prefs，
-  /// 那部分旧进度在 prefs 里缺失。故 prefs 无记录时回退查 `VideoBooks.lastPositionMs`
-  /// （旧数据无独立时间戳记 0），与 prefs 经 [resolvePositionLww] 取较新——既能读
-  /// 出旧本机播放进度（client 跨设备恢复），又不让无时间戳的旧值盖过更新的 prefs 进度。
-  @override
-  Future<({int positionMs, int updatedAtMs})> getVideoPosition(
-    String id, {
-    int episodeIndex = 0,
-  }) async {
-    final int prefsPos = await _db.getPrefTyped<int>(
-        videoRemotePositionEpisodePrefKey(id, episodeIndex), 0);
-    final int prefsAt = await _db.getPrefTyped<int>(
-        videoRemotePositionEpisodeAtPrefKey(id, episodeIndex), 0);
-    // 旧 host 本机播放只写 VideoBooks.lastPositionMs（整书一个值，无按集语义）；只在
-    // episodeIndex<=0（当前集 / 单视频）回退它，避免给某集错配整书的旧进度。
-    final VideoBookRow? row =
-        episodeIndex <= 0 ? await _db.getVideoBookByBookUid(id) : null;
-    final int rowPos = row?.lastPositionMs ?? 0;
-    // BUG-996：lastPositionMs 列无时间戳，此前硬编码 remoteUpdatedAtMs:0，使 host 的真
-    // 进度在跨设备 LWW 里恒输给任何带 now 戳的本地断点（client 一旦碰过就再也拉不回
-    // host 的桌面新进度）。用 importedAt 作「进度至少和导入一样旧」的可辩护下界戳——
-    // client 真更近才看过仍会赢（语义可接受），但 client 无有效断点时 host 能续上。
-    final int rowAt = rowPos > 0 ? (row?.importedAt ?? 0) : 0;
-    return resolvePositionLww(
-      localPositionMs: prefsPos,
-      localUpdatedAtMs: prefsAt,
-      remotePositionMs: rowPos,
-      remoteUpdatedAtMs: rowAt,
-    );
-  }
-
-  /// 把 client 上报的 [id] 视频断点写入 host（TODO-653）。
-  ///
-  /// 冲突解决「取较新时间戳」（[resolvePositionLww]）：仅当 [updatedAtMs] 严格
-  /// 新于 host 已存时间戳才覆盖，避免旧设备滞后上报回退新进度。负位置 clamp 0。
-  @override
-  Future<void> putVideoPosition(
-    String id,
-    int positionMs,
-    int updatedAtMs, {
-    int episodeIndex = 0,
-  }) async {
-    final ({int positionMs, int updatedAtMs}) current =
-        await getVideoPosition(id, episodeIndex: episodeIndex);
-    final ({int positionMs, int updatedAtMs}) winner = resolvePositionLww(
-      localPositionMs: current.positionMs,
-      localUpdatedAtMs: current.updatedAtMs,
-      remotePositionMs: positionMs < 0 ? 0 : positionMs,
-      remoteUpdatedAtMs: updatedAtMs,
-    );
-    if (winner.updatedAtMs == current.updatedAtMs &&
-        winner.positionMs == current.positionMs) {
-      return; // host 已存更新或相等，no-op。
-    }
-    await _db.setPrefTyped<int>(
-        videoRemotePositionEpisodePrefKey(id, episodeIndex), winner.positionMs);
-    await _db.setPrefTyped<int>(
-        videoRemotePositionEpisodeAtPrefKey(id, episodeIndex),
-        winner.updatedAtMs);
-    // BUG-1731：prefs 键空间只被「下发清单给子端」消费，host 自己的 UI（继续观看
-    // / 下一集 / 合集续播锚点）读的是 VideoBooks.lastPositionMs / lastPlayedAt。
-    // 子端上报只写 prefs 会让 host 端 UI 永远看不到对端进度——胜者来自对端时镜像
-    // 写行，与 client 侧 sync_orchestrator 的 writeBackLocal 同纪律。playedAt 用
-    // **对端的** updatedAtMs（绝不是 now）：传 now 会把对方三天前看的冒充成本机刚
-    // 看的，钉死合集续播锚点（BUG-1542）。仅 episodeIndex<=0（合集每集一行 / 单
-    // 视频）镜像；episodeIndex>0 是 host-playlist 单行多集形态，行级
-    // lastPositionMs 无按集语义，写它会把某一集的进度错配成整行进度。
-    if (episodeIndex <= 0 && await _db.getVideoBookByBookUid(id) != null) {
-      await _db.updateVideoBookPosition(id, winner.positionMs,
-          playedAt: winner.updatedAtMs);
-    }
-  }
-
-  /// [id] 视频的主归属合集行（无归属 / 未知 id 返回 null）。系列级播放偏好
-  /// （`subtitleDelayMs`）解析用，与清单侧 [_primaryCollectionData] 同折叠语义
-  /// （最小 collectionId 主归属）。两次轻查询（主归属映射 + 单行取合集）。
-  Future<MediaCollectionRow?> _primaryVideoCollectionRow(String id) async {
-    final Map<String, int> primaryByEntry =
-        await _db.getPrimaryCollectionIdByEntry();
-    final int? cid = primaryByEntry[MediaKind.video.compositeKey(id)];
-    if (cid == null) return null;
-    return _db.getMediaCollectionById(cid);
-  }
-
-  /// prefs 里的可空字符串字段读数：缺失/空串 → null（「未设」与「显式清除」由
-  /// at 键区分：at>0 且值空 = 显式清除）。
-  Future<String?> _readNullableStringPref(String key) async {
-    final String raw = await _db.getPrefTyped<String>(key, '');
-    return raw.isEmpty ? null : raw;
-  }
-
-  /// 读 host 端视频 [id] 的播放偏好带戳状态（[VideoPlaybackSyncHost]）。
-  ///
-  /// 无戳基底 = 本机播放读的同一决议（系列级 ?? row，[effectiveSeriesDelayMs] 族），
-  /// 与带戳 prefs 逐字段「严格较新者胜」合并：旧数据无 prefs 时行为与本机播放完全
-  /// 一致（向后兼容）。
-  @override
-  Future<VideoPlaybackSyncState> getVideoPlayback(String id) async {
-    final VideoBookRow? row = await _db.getVideoBookByBookUid(id);
-    final MediaCollectionRow? col =
-        row == null ? null : await _primaryVideoCollectionRow(id);
-    final VideoPlaybackSyncState base = VideoPlaybackSyncState(
-      delayMs: effectiveSeriesDelayMs(col?.subtitleDelayMs, row?.delayMs ?? 0),
-      audioTrackId:
-          effectiveSeriesAudioTrackId(col?.audioTrackId, row?.audioTrackId),
-      secondarySubtitleSource: row?.secondarySubtitleSource,
-      secondaryDelayMs: effectiveSeriesSecondaryDelayMs(
-          col?.secondarySubtitleDelayMs, row?.secondaryDelayMs),
-    );
-    final VideoPlaybackSyncState stamped = VideoPlaybackSyncState(
-      delayMs: await _db.getPrefTyped<int>(videoRemoteDelayPrefKey(id), 0),
-      delayAt: await _db.getPrefTyped<int>(videoRemoteDelayAtPrefKey(id), 0),
-      audioTrackId:
-          await _readNullableStringPref(videoRemoteAudioTrackPrefKey(id)),
-      audioTrackAt:
-          await _db.getPrefTyped<int>(videoRemoteAudioTrackAtPrefKey(id), 0),
-      secondarySubtitleSource: await _readNullableStringPref(
-          videoRemoteSecondarySubtitlePrefKey(id)),
-      secondarySubtitleAt: await _db.getPrefTyped<int>(
-          videoRemoteSecondarySubtitleAtPrefKey(id), 0),
-      secondaryDelayMs: int.tryParse(await _db.getPrefTyped<String>(
-          videoRemoteSecondaryDelayPrefKey(id), '')),
-      secondaryDelayAt: await _db.getPrefTyped<int>(
-          videoRemoteSecondaryDelayAtPrefKey(id), 0),
-    );
-    return VideoPlaybackSyncState.merge(base, stamped);
-  }
-
-  /// 把 client 上报的播放偏好合并进 host（[VideoPlaybackSyncHost]）。
-  ///
-  /// 存在性闸门：host 无该 id 的 VideoBooks 行 → no-op（防任意 id 写脏 prefs，与
-  /// [putVideoPosition] 同语义）。调轴类字段越界 clamp
-  /// ±[kVideoSubtitleDelayLimitMs]；各字段时间戳截到 host 当前时刻 + 5 分钟时钟
-  /// 偏差余量——未来戳会永久锁死后续所有端的正常覆盖。胜出字段同时写穿
-  /// row/系列级列，使 host 本机播放立即跟随（只写 row 会被非 null 系列级值遮蔽）。
-  @override
-  Future<void> putVideoPlayback(
-      String id, VideoPlaybackSyncState incoming) async {
-    if (incoming.isEmpty) return;
-    if (await _db.getVideoBookByBookUid(id) == null) return;
-    final int nowCapMs = DateTime.now().millisecondsSinceEpoch + 5 * 60 * 1000;
-    int capAt(int at) => at > nowCapMs ? nowCapMs : at;
-    final VideoPlaybackSyncState capped = VideoPlaybackSyncState(
-      delayMs: incoming.delayMs
-          .clamp(-kVideoSubtitleDelayLimitMs, kVideoSubtitleDelayLimitMs),
-      delayAt: capAt(incoming.delayAt),
-      audioTrackId: incoming.audioTrackId,
-      audioTrackAt: capAt(incoming.audioTrackAt),
-      secondarySubtitleSource: incoming.secondarySubtitleSource,
-      secondarySubtitleAt: capAt(incoming.secondarySubtitleAt),
-      secondaryDelayMs: incoming.secondaryDelayMs
-          ?.clamp(-kVideoSubtitleDelayLimitMs, kVideoSubtitleDelayLimitMs),
-      secondaryDelayAt: capAt(incoming.secondaryDelayAt),
-    );
-    final VideoPlaybackSyncState held = await getVideoPlayback(id);
-    final VideoPlaybackSyncState merged =
-        VideoPlaybackSyncState.merge(held, capped);
-    if (merged == held) return; // host 已存更新或相等，no-op。
-    final MediaCollectionRow? col = await _primaryVideoCollectionRow(id);
-
-    if (merged.delayAt > 0 && merged.delayAt != held.delayAt) {
-      await _db.setPrefTyped<int>(videoRemoteDelayPrefKey(id), merged.delayMs);
-      await _db.setPrefTyped<int>(
-          videoRemoteDelayAtPrefKey(id), merged.delayAt);
-      await _db.updateVideoBookDelayMs(id, merged.delayMs);
-      if (col != null) {
-        await _db.updateMediaCollectionSubtitleDelayMs(col.id, merged.delayMs);
-      }
-    }
-    if (merged.audioTrackAt > 0 && merged.audioTrackAt != held.audioTrackAt) {
-      await _db.setPrefTyped<String>(
-          videoRemoteAudioTrackPrefKey(id), merged.audioTrackId ?? '');
-      await _db.setPrefTyped<int>(
-          videoRemoteAudioTrackAtPrefKey(id), merged.audioTrackAt);
-      await _db.updateVideoBookAudioTrackId(id, merged.audioTrackId);
-      if (col != null) {
-        await _db.updateMediaCollectionAudioTrackId(
-            col.id, merged.audioTrackId);
-      }
-    }
-    if (merged.secondarySubtitleAt > 0 &&
-        merged.secondarySubtitleAt != held.secondarySubtitleAt) {
-      await _db.setPrefTyped<String>(videoRemoteSecondarySubtitlePrefKey(id),
-          merged.secondarySubtitleSource ?? '');
-      await _db.setPrefTyped<int>(videoRemoteSecondarySubtitleAtPrefKey(id),
-          merged.secondarySubtitleAt);
-      await _db.updateVideoBookSecondarySubtitleSource(
-          id, merged.secondarySubtitleSource);
-    }
-    if (merged.secondaryDelayAt > 0 &&
-        merged.secondaryDelayAt != held.secondaryDelayAt) {
-      await _db.setPrefTyped<String>(videoRemoteSecondaryDelayPrefKey(id),
-          merged.secondaryDelayMs?.toString() ?? '');
-      await _db.setPrefTyped<int>(
-          videoRemoteSecondaryDelayAtPrefKey(id), merged.secondaryDelayAt);
-      await _db.updateVideoBookSecondaryDelayMs(id, merged.secondaryDelayMs);
-      if (col != null) {
-        await _db.updateMediaCollectionSecondarySubtitleDelayMs(
-            col.id, merged.secondaryDelayMs);
-      }
-    }
-  }
-
-  /// 廉价判断 host 库是否已存在 bookUid 为 [id] 的视频（一次 DB 查询）。
-  @override
-  Future<bool> videoExists(String id) async {
-    _assertSafeVideoId(id);
-    return (await _db.getVideoBookByBookUid(id)) != null;
-  }
-
-  /// 从 host 视频库删除 bookUid 为 [id] 的视频（[VideoDeletionHost]）。
-  ///
-  /// 与 host 用户在自己视频库长按删除同语义：repository 的完整删除 operation 负责
-  /// DB 行 + 字幕 cue + 合集引用 + 删除墓碑及 app-owned 配图/字幕回收；host 再回收
-  /// 自己接收上传时创建的视频副本。用户自己导入的原始视频文件绝不删除。
-  @override
-  Future<void> deleteVideo(String id) async {
-    _assertSafeVideoId(id);
-    final VideoScrapeOperationLease? lease =
-        VideoScrapeOperationGate.tryEnterOperation();
-    if (lease == null) throw StateError('视频刮削资料正在清理');
-    try {
-      await _runExclusive(
-        () => VideoCoverMutationGate.runExclusive(() async {
-          final VideoBookRow? row = await _db.getVideoBookByBookUid(id);
-          if (row == null) return; // 幂等：不存在则静默跳过
-          final bool deleted =
-              await VideoBookRepository(_db).deleteVideoBookAndReclaimAssets(
-            id,
-            scope: DeleteScope.syncEverywhere,
-            compactDatabase: false,
-          );
-          if (!deleted) return;
-          await _deleteUploadedVideoCopy(row);
-        }),
-      );
-    } finally {
-      lease.release();
-    }
-  }
-
-  /// 删除 client 上传副本目录——当且仅当该行的 `videoPath` 确实落在本 host 的
-  /// `<uploadedVideoRoot>/<safeUid>/` 之内。
-  ///
-  /// host 用户自己导入的原片不在这个目录下，所以这条 [p.isWithin] 判据就是
-  /// 「这些字节是 app 自己搬进来的」的证明；判据不成立时一个字节都不动。
-  Future<void> _deleteUploadedVideoCopy(VideoBookRow row) async {
-    final Directory? root = _uploadedVideoRoot;
-    if (root == null) return;
-    final Directory owned =
-        Directory(p.join(root.path, _sanitizeVideoIdForPath(row.bookUid)));
-    if (!owned.existsSync()) return;
-    if (!p.isWithin(owned.path, row.videoPath)) return;
-    try {
-      await owned.delete(recursive: true);
-    } catch (_) {
-      // best-effort：目录被占用等失败不影响 DB 已删。
-    }
-  }
-
-  /// 接收 client 上传的单文件视频并注册进 host 视频库（client→host live push）。
-  ///
-  /// 落盘目录按 [id] 确定（`<uploadedVideoRoot>/<safeUid>/`），故重复上传同一视频
-  /// 覆盖同一副本、不留孤儿；`upsertVideoBook` 幂等按 bookUid 覆盖同一行。封面 best-effort
-  /// 抽取，与建行解耦（绝不挡上传落库）。
-  @override
-  Future<void> importVideo(
-    File videoFile, {
-    required String id,
-    required String title,
-    String? originalFileName,
-  }) async {
-    _assertSafeVideoId(id);
-    final Directory? root = _uploadedVideoRoot;
-    if (root == null) {
-      throw UnsupportedError(
-        'importVideo requires uploadedVideoRoot to be provided',
-      );
-    }
-    await _runExclusive(() async {
-      final String safeUid = _sanitizeVideoIdForPath(id);
-      final Directory destDir = Directory(p.join(root.path, safeUid));
-      destDir.createSync(recursive: true);
-      final File dest = File(
-          p.join(destDir.path, _uploadedVideoFileName(originalFileName, id)));
-      await _moveFileInto(videoFile, dest);
-      await _db.upsertVideoBook(VideoBooksCompanion(
-        bookUid: Value(id),
-        title: Value(title),
-        videoPath: Value(dest.path),
-        // 无外挂字幕上传：回退内嵌默认轨（与 client 下载无字幕分支一致）。
-        embeddedSubtitleTrack: const Value<int?>(0),
-        importedAt: Value(DateTime.now().millisecondsSinceEpoch),
-      ));
-    });
-    // 封面 best-effort，与建行解耦：抽帧走 ffmpeg 慢，失败留空占位（移动端无 ffmpeg
-    // 返 null），绝不让封面失败使整个上传报错。
-    final Future<String?> Function(
-        {required String videoPath,
-        required String bookUid})? extractor = _extractVideoCover;
-    if (extractor != null) {
-      final VideoScrapeOperationLease? lease =
-          VideoScrapeOperationGate.tryEnterOperation();
-      if (lease != null) {
-        try {
-          await VideoCoverMutationGate.runExclusive(() async {
-            final VideoBookRow? row = await _db.getVideoBookByBookUid(id);
-            if (row == null) return;
-            final CoverMetaStore store = CoverMetaStore(
-              _videoCoversDirectory ?? await VideoStorage.coversDir(),
-            );
-            if (!await store.allowsAutoFrameWrite(id)) return;
-            final String? coverPath = await extractor(
-              videoPath: row.videoPath,
-              bookUid: id,
-            );
-            if (coverPath != null && coverPath.isNotEmpty) {
-              await _db.updateVideoBookCover(id, coverPath);
-              final bool committed = await store.markAutoFrameAfterWrite(id);
-              if (!committed) {
-                ErrorLogService.instance.log(
-                  'sync.videoCover.provenanceConflict',
-                  StateError('封面来源在自动抽帧期间发生变化: $id'),
-                  StackTrace.current,
-                );
-              }
-            }
-          });
-        } catch (error, stack) {
-          // best-effort：封面失败不影响上传成功。
-          ErrorLogService.instance.log(
-            'sync.videoCover.backfill',
-            error,
-            stack,
-          );
-        } finally {
-          lease.release();
-        }
-      }
-    }
-  }
-
-  /// 接收 client 上传的视频外挂字幕并落到视频同目录（BUG-964，client→host live push）。
-  ///
-  /// 落盘名 = `<host 视频文件 stem><suffix>`，与 [resolveVideoSubtitle] 的同 stem
-  /// 匹配规则天然一致；[suffix] 经 [isSidecarSubtitleSuffix] 白名单校验（拒路径
-  /// 分隔符/穿越）。落位后按 host 学习语言重解析首选 sidecar（多字幕推送顺序无关、
-  /// 结果收敛），镜像 client 下载路径（home_video_page `_registerDownloadedVideo`）
-  /// 的行语义：`subtitleSource`/`subtitleFormat` 指向首选 sidecar、
-  /// `embeddedSubtitleTrack=null`（播放走外挂）、解析 cue 落库（坏字幕 best-effort
-  /// 跳过，不挡文件落位——host 仍能把字节原样转发给其它 client）。
-  @override
-  Future<void> importVideoSubtitle(
-    File subtitleFile, {
-    required String id,
-    required String suffix,
-  }) async {
-    _assertSafeVideoId(id);
-    if (!isSidecarSubtitleSuffix(suffix)) {
-      throw ArgumentError.value(suffix, 'suffix', 'unsafe subtitle suffix');
-    }
-    await _runExclusive(() async {
-      final VideoBookRow? row = await _db.getVideoBookByBookUid(id);
-      if (row == null) throw StateError('unknown video: $id');
-      final String videoPath = row.videoPath;
-      final String lower = videoPath.toLowerCase();
-      if (videoPath.isEmpty ||
-          lower.startsWith('http://') ||
-          lower.startsWith('https://')) {
-        throw StateError('video has no local file: $id');
-      }
-      final File dest = File(p.join(p.dirname(videoPath),
-          '${p.basenameWithoutExtension(videoPath)}$suffix'));
-      await _moveFileInto(subtitleFile, dest);
-      final String? preferred =
-          findSidecarSubtitle(videoPath, langCode: _videoSubtitleLangCode);
-      if (preferred == null) return; // 防御：刚落位的 dest 本身就是候选。
-      final String ext =
-          p.extension(preferred).replaceFirst('.', '').toLowerCase();
-      List<AudioCue> cues = const <AudioCue>[];
-      try {
-        cues = parseSubtitleCues(
-          content: await readTextWithEncoding(File(preferred)),
-          format: ext,
-          bookUid: id,
-        );
-      } catch (_) {
-        // best-effort：解析失败不挡字幕文件落位。
-      }
-      await _db.upsertVideoBook(VideoBooksCompanion(
-        bookUid: Value(id),
-        title: Value(row.title),
-        videoPath: Value(row.videoPath),
-        subtitleSource: Value<String?>(preferred),
-        subtitleFormat: Value<String?>(ext),
-        embeddedSubtitleTrack: const Value<int?>(null),
-      ));
-      if (cues.isNotEmpty) {
-        await _db.replaceCuesForBook(
-            id, cues.map(AudioCue.toCompanion).toList());
-      }
-    });
-  }
-
-  /// 校验视频 id 不含路径穿越字符（`..` / `\`）。id 允许 `/`（bookUid 形如
-  /// `video/xxx`），落盘前经 [_sanitizeVideoIdForPath] 压平。
-  static void _assertSafeVideoId(String id) {
-    if (id.isEmpty || id.contains('..') || id.contains('\\')) {
-      throw ArgumentError.value(id, 'id', 'unsafe video id');
-    }
-  }
-
-  /// 把视频 id 压成单层安全目录名（`/` → `_`，其余非白名单字符 → `_`）。
-  static String _sanitizeVideoIdForPath(String id) =>
-      id.replaceAll(RegExp(r'[^A-Za-z0-9._-]'), '_');
-
-  /// 上传视频落盘文件名：优先保留上传方原始 basename 的扩展名（media_kit 依赖扩展名
-  /// 判容器）；basename 缺失/不安全时回退 `<safeUid>.mp4`。
-  static String _uploadedVideoFileName(String? originalFileName, String id) {
-    if (originalFileName != null && originalFileName.isNotEmpty) {
-      final String base = p.basename(originalFileName);
-      final String safe = base.replaceAll(RegExp(r'[/\\]'), '_');
-      if (safe.isNotEmpty &&
-          !safe.contains('..') &&
-          p.extension(safe).isNotEmpty) {
-        return safe;
-      }
-    }
-    return '${_sanitizeVideoIdForPath(id)}.mp4';
-  }
-
-  /// 把 [src] 搬进 [dest]（同卷 rename 最快；跨卷 rename 失败回退 copy + delete）。
-  static Future<void> _moveFileInto(File src, File dest) async {
-    try {
-      await src.rename(dest.path);
-    } on FileSystemException {
-      await src.copy(dest.path);
-      try {
-        await src.delete();
-      } catch (_) {
-        // best-effort：源临时文件删除失败不影响落库（上层临时目录整体清理）。
-      }
-    }
-  }
-
-  // ── 聚合（统计 + 收藏，TODO-1056 phase C）────────────────────────────────────
-
-  /// 读 host 端当前聚合快照。直接复用云后端 phase B 的
-  /// [AggregateSyncService.materializeLocalSnapshot]（同一 DB 读取逻辑），保证互联
-  /// 与云通道 materialize 结果字节等价、无第二套实现。
-  @override
-  Future<AggregateSnapshot> getAggregateSnapshot() async {
-    return AggregateSyncService(_db).materializeLocalSnapshot();
-  }
-
-  /// 把 client 上报的聚合快照折叠进 host DB。用
-  /// [AggregateSyncService.foldIntoLocal]（先 materialize host 自己 → MAX / 并集
-  /// 合并 incoming → apply），保证 host 侧也满足 never-shrinks：client 上报的某字段
-  /// 即便小于 host 当前值（并发 / GET 后 host 又涨），MAX 折叠让 host 值不被缩小；
-  /// 幂等（重复 apply 同一快照不变）；删除不跨端传播。经 [_runExclusive] 与其它库
-  /// 变动串行，避免与 host 本机写统计/收藏竞态。
-  @override
-  Future<void> applyAggregateSnapshot(AggregateSnapshot snapshot) async {
-    await _runExclusive(
-      () => AggregateSyncService(_db).foldIntoLocal(snapshot),
-    );
-  }
-
-  // ── 合集清单（多端库联合视图 §2.3 任务5.2）──────────────────────────────────
-
-  /// 读 host 合集全量快照清单。直接复用云后端同一 [loadLocalCollectionManifest]（同一
-  /// DB 读取逻辑），保证互联与云通道 materialize 结果字节等价、无第二套实现。
-  @override
-  Future<CollectionManifest> getCollectionManifest() =>
-      loadLocalCollectionManifest(_db);
-
-  /// 把 client 上报的合集清单并入 host DB 并返回合并后清单。
-  ///
-  /// 与云后端 orchestrator [SyncOrchestrator.syncCollections] 的核心完全同构，仅
-  /// 通道不同：`CollectionSyncEngine.merge`（host 自身 `sync_collections_baseline_ms`
-  /// 因果基线）→ [applyCollectionLocalChanges] 把本地变更集落 host DB → 推进 host
-  /// 基线 → 返回合并后清单（client 端再拿它重跑引擎收敛，双端同一并集）。
-  ///
-  /// 基线的角色：区分「未见过的移出/删除墓碑」（新闻 → 生效）与「本端已裁决过、
-  /// 成员/合集仍在即代表之后重加/重建」（旧闻 → 活胜）。host 每次合并成功后推进基线，
-  /// 使已应用的墓碑成为「旧闻」，日后本端或对端重加时不被旧墓碑再删（收敛正确性依赖
-  /// 此推进，见 collection_sync_engine 注释）。
-  ///
-  /// 经 [_runExclusive] 与其它库变动串行：读清单→合并→落库→推基线整体互斥，
-  /// 避免与 host 本机合集编辑竞态。重放同一清单幂等（应用端按目标态调和）。
-  @override
-  Future<CollectionManifest> mergeCollectionManifest(
-    CollectionManifest incoming,
-  ) async {
-    late CollectionManifest merged;
-    await _runExclusive(() async {
-      final CollectionManifest local = await loadLocalCollectionManifest(_db);
-      final SyncRepository repo = SyncRepository(_db);
-      // BUG-1579：host 收 POST 走**自己那本账**（[SyncChannelScope.host]）。它与
-      // 本机作为 client 跑的云/互联通道是三条独立的因果轴：共用一个键时，client
-      // 通道刚推进的基线会把对端 POST 来的移出墓碑判成旧闻 → 引擎按「活胜」撤销
-      // → 再回传给对端，用户的移出被自己另一条通道悄悄撤销。
-      final int baseline =
-          await repo.getCollectionsSyncBaselineMs(SyncChannelScope.host);
-      final CollectionSyncOutcome outcome = CollectionSyncEngine.merge(
-        local: local,
-        remote: incoming,
-        lastSyncedAtMs: baseline,
-      );
-      await applyCollectionLocalChanges(_db, outcome.changes);
-      await repo.setCollectionsSyncBaselineMs(
-          SyncChannelScope.host, DateTime.now().millisecondsSinceEpoch);
-      merged = outcome.merged;
-    });
-    return merged;
-  }
-
-  @override
-  Future<List<({String mediaType, String itemKey, int deletedAt})>>
-      listDeletionTombstones() async {
-    final List<SyncDeletionTombstoneRow> rows =
-        await _db.getSyncDeletionTombstones();
-    return <({String mediaType, String itemKey, int deletedAt})>[
-      for (final SyncDeletionTombstoneRow r in rows)
-        (mediaType: r.mediaType, itemKey: r.itemKey, deletedAt: r.deletedAt),
-    ];
   }
 }
