+++ fushi/test/sync/book_rename_lww_test.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:fushi/src/media/media_source.dart';
 import 'package:fushi/src/media/sources/reader_fushi_source.dart';
-import 'package:fushi/src/sync/app_model_library_host_service.dart';
+import 'package:fushi/src/sync/local_library_host_service.dart';
 import 'package:fushi/src/sync/backup_merge_engine.dart';
 import 'package:fushi/src/sync/fushi_library_host_service.dart';
 import 'package:fushi/src/sync/interconnect_sync_backend.dart';
@@ -34,7 +34,7 @@ import 'temp_dir_cleanup.dart';
 ///    → 「平局保留本机」两例转红（旧对端把本机改名覆盖掉）；
 ///  - 同函数 `updatedAt: Value<int>(updatedAt)` 改成写 `now`
 ///    → 「戳落成对端的戳」转红；
-///  - `app_model_library_host_service.dart` 的 `displayTitleAt:
+///  - `local_library_host_service.dart` 的 `displayTitleAt:
 ///    overrideTitles[r.bookKey]?.updatedAt ?? 0` 改成恒 `0`
 ///    → 「host 清单下发改名时刻」转红；
 ///  - `fushi_library_host_service.dart` wire 键 `'displayTitleAt'` 改名成
@@ -67,11 +67,11 @@ void main() {
     return db;
   }
 
-  AppModelLibraryHostService buildHost(
+  LocalLibraryHostService buildHost(
     FushiDatabase db, {
     Future<String?> Function(File)? importBookFromFile,
   }) =>
-      AppModelLibraryHostService(
+      LocalLibraryHostService(
         db: db,
         dictionaryResourceRoot: Directory.systemTemp,
         packages: SyncAssetPackageService(db: db),
@@ -404,7 +404,7 @@ void main() {
     final ReaderFushiSource source = await bindSource(db);
 
     // fake importer：落库并返回**真实** bookKey（重名会带后缀，与 title 不同）。
-    final AppModelLibraryHostService host = buildHost(
+    final LocalLibraryHostService host = buildHost(
       db,
       importBookFromFile: (File f) async {
         const String bookKey = '原始書名 (2)';
@@ -437,7 +437,7 @@ void main() {
   test('host 收到旧 client 的 push（无 header）：行为与本轮之前逐字相同', () async {
     final FushiDatabase db = await openDb('bug1503_host_old_');
     final ReaderFushiSource source = await bindSource(db);
-    final AppModelLibraryHostService host = buildHost(
+    final LocalLibraryHostService host = buildHost(
       db,
       importBookFromFile: (File f) async {
         await db.insertEpubBook(book('書', '書'));
