+++ fushi/test/sync/remote_info_tags_test.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:drift/drift.dart' hide isNull, isNotNull;
 import 'package:drift/native.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:fushi/src/sync/app_model_library_host_service.dart';
+import 'package:fushi/src/sync/local_library_host_service.dart';
 import 'package:fushi/src/sync/fushi_library_host_service.dart';
 import 'package:fushi/src/sync/sync_asset_package_service.dart';
 import 'package:fushi_core/fushi_core.dart';
@@ -19,8 +19,8 @@ void main() {
     return db;
   }
 
-  AppModelLibraryHostService buildSvc(FushiDatabase db) =>
-      AppModelLibraryHostService(
+  LocalLibraryHostService buildSvc(FushiDatabase db) =>
+      LocalLibraryHostService(
         db: db,
         dictionaryResourceRoot: Directory.systemTemp,
         packages: SyncAssetPackageService(db: db),
