+++ fushi/test/sync/sync_orchestrator_live_audio_test.dart
@@ -21,7 +21,7 @@ import 'package:fushi/src/models/local_audio_manager.dart'
     show LocalAudioDbEntry;
 import 'package:fushi/src/models/local_audio_source_pref.dart'
     show LocalAudioSourcePref;
-import 'package:fushi/src/sync/app_model_library_host_service.dart';
+import 'package:fushi/src/sync/local_library_host_service.dart';
 import 'package:fushi/src/sync/interconnect_sync_backend.dart';
 import 'package:fushi/src/sync/fushi_sync_server.dart';
 import 'package:fushi/src/sync/sync_asset_package_service.dart';
@@ -330,7 +330,7 @@ void main() {
         ),
       ];
 
-      final AppModelLibraryHostService libSvc = AppModelLibraryHostService(
+      final LocalLibraryHostService libSvc = LocalLibraryHostService(
         db: hostDb,
         dictionaryResourceRoot: Directory(work.path),
         packages: SyncAssetPackageService(db: hostDb),
@@ -529,7 +529,7 @@ void main() {
         ),
       );
 
-      final AppModelLibraryHostService libSvc = AppModelLibraryHostService(
+      final LocalLibraryHostService libSvc = LocalLibraryHostService(
         db: hostDb,
         dictionaryResourceRoot: Directory(work.path),
         packages: SyncAssetPackageService(db: hostDb),
@@ -746,7 +746,7 @@ void main() {
         extractDir: p.join(work.path, 'host_b2_text_extract'),
       );
 
-      final AppModelLibraryHostService libSvc = AppModelLibraryHostService(
+      final LocalLibraryHostService libSvc = LocalLibraryHostService(
         db: hostDb,
         dictionaryResourceRoot: Directory(work.path),
         packages: SyncAssetPackageService(db: hostDb),
@@ -853,7 +853,7 @@ void main() {
     setUp(() async {
       hostDb = _memDb();
       // host 无需真实数据，只要 server 能响应
-      final AppModelLibraryHostService libSvc = AppModelLibraryHostService(
+      final LocalLibraryHostService libSvc = LocalLibraryHostService(
         db: hostDb,
         dictionaryResourceRoot: Directory(work.path),
         packages: SyncAssetPackageService(db: hostDb),
@@ -987,7 +987,7 @@ void main() {
       final Directory hostAudioRoot =
           Directory(p.join(work.path, 'host_e_audio_root'))
             ..createSync(recursive: true);
-      final AppModelLibraryHostService libSvc = AppModelLibraryHostService(
+      final LocalLibraryHostService libSvc = LocalLibraryHostService(
         db: hostDb,
         dictionaryResourceRoot: Directory(work.path),
         packages: SyncAssetPackageService(db: hostDb),
