+++ fushi/test/sync/sync_orchestrator_live_book_test.dart
@@ -18,7 +18,7 @@ import 'dart:io';
 import 'package:drift/drift.dart' hide isNull, isNotNull;
 import 'package:drift/native.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:fushi/src/sync/app_model_library_host_service.dart';
+import 'package:fushi/src/sync/local_library_host_service.dart';
 import 'package:fushi/src/sync/interconnect_sync_backend.dart';
 import 'package:fushi/src/sync/fushi_sync_server.dart';
 import 'package:fushi/src/sync/sync_asset_package_service.dart';
@@ -272,7 +272,7 @@ void main() {
       final String hostExtract = p.join(work.path, 'host_extract_Y');
       await _seedBook(db: hostDb, title: 'BookY', extractDir: hostExtract);
 
-      final AppModelLibraryHostService libSvc = AppModelLibraryHostService(
+      final LocalLibraryHostService libSvc = LocalLibraryHostService(
         db: hostDb,
         dictionaryResourceRoot: Directory(work.path),
         packages: SyncAssetPackageService(db: hostDb),
@@ -443,7 +443,7 @@ void main() {
       final String hostExtract = p.join(work.path, 'host_extract_Y_b');
       await _seedBook(db: hostDb, title: 'BookY', extractDir: hostExtract);
 
-      final AppModelLibraryHostService libSvc = AppModelLibraryHostService(
+      final LocalLibraryHostService libSvc = LocalLibraryHostService(
         db: hostDb,
         dictionaryResourceRoot: Directory(work.path),
         packages: SyncAssetPackageService(db: hostDb),
