+++ fushi/test/sync/sync_orchestrator_live_video_test.dart
@@ -12,7 +12,7 @@ import 'dart:io';
 import 'package:drift/drift.dart' hide isNull, isNotNull;
 import 'package:drift/native.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:fushi/src/sync/app_model_library_host_service.dart';
+import 'package:fushi/src/sync/local_library_host_service.dart';
 import 'package:fushi/src/sync/interconnect_sync_backend.dart';
 import 'package:fushi/src/sync/fushi_sync_server.dart';
 import 'package:fushi/src/sync/sync_asset_package_service.dart';
@@ -74,7 +74,7 @@ void main() {
     hostDb = _memDb();
     hostUploads = Directory(p.join(work.path, 'host_uploads'))
       ..createSync(recursive: true);
-    final AppModelLibraryHostService libSvc = AppModelLibraryHostService(
+    final LocalLibraryHostService libSvc = LocalLibraryHostService(
       db: hostDb,
       dictionaryResourceRoot: Directory(work.path),
       packages: SyncAssetPackageService(db: hostDb),
